% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{as.character.pugi_xml}
\alias{as.character.pugi_xml}
\title{as.character pugi_xml}
\usage{
\method{as.character}{pugi_xml}(x, ...)
}
\arguments{
\item{x}{a class pugi_xml object}

\item{...}{additional arguments see \code{\link[=print.pugi_xml]{print.pugi_xml()}}}
}
\description{
as.character pugi_xml
}
\examples{
  # a pointer
  x <- read_xml("<a><b/></a>")
  as.character(x)
  as.character(x, raw = FALSE)
}
