% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illegal-characters.R
\name{clean_worksheet_name}
\alias{clean_worksheet_name}
\title{Clean worksheet name}
\usage{
clean_worksheet_name(x)
}
\arguments{
\item{x}{A vector, coherced to \code{character}}
}
\value{
x with bad characters removed
}
\description{
Clean worksheet name
}
