% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{cloneSheetStyle}
\alias{cloneSheetStyle}
\title{clone sheets style}
\usage{
cloneSheetStyle(wb, from_sheet, to_sheet)
}
\arguments{
\item{wb}{workbook}

\item{from_sheet}{sheet we select the style from}

\item{to_sheet}{sheet we apply the style from}
}
\description{
clone sheets style
}
