% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{cloneWorksheet}
\alias{cloneWorksheet}
\alias{wb_clone_worksheet}
\title{Clone a worksheet to a workbook}
\usage{
wb_clone_worksheet(wb, sheetName, clonedSheet)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet}

\item{sheetName}{A name for the new worksheet}

\item{clonedSheet}{The name of the existing worksheet to be cloned.}
}
\value{
The \code{wb} object
}
\description{
Clone a worksheet to a Workbook object
}
\examples{
## Create a new workbook
wb <- wb_workbook("Fred")

## Add 3 worksheets
wb$addWorksheet("Sheet 1")
wb_clone_worksheet(wb, "Sheet 2", clonedSheet = "Sheet 1")

## Save workbook
\dontrun{
wb_save(wb, "cloneWorksheetExample.xlsx", overwrite = TRUE)
}
}
