% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{createComment}
\alias{createComment}
\alias{writeComment}
\alias{removeComment}
\title{Create, write and remove comments}
\usage{
createComment(
  text,
  author = Sys.info()[["user"]],
  style = NULL,
  visible = TRUE,
  width = 2,
  height = 4
)

writeComment(wb, sheet, col, row, comment, xy = NULL)

removeComment(wb, sheet, col, row, gridExpand = TRUE)
}
\arguments{
\item{text}{Comment text. Character vector.}

\item{author}{Author of comment. Character vector of length 1}

\item{style}{A Style object or list of style objects the same length as comment vector. See \code{\link[=createStyle]{createStyle()}}.}

\item{visible}{TRUE or FALSE. Is comment visible.}

\item{width}{Textbox integer width in number of cells}

\item{height}{Textbox integer height in number of cells}

\item{wb}{A workbook object}

\item{sheet}{A vector of names or indices of worksheets}

\item{col}{Column a column number of letter. For \code{removeComment} this can be a range.}

\item{row}{A row number. For \code{removeComment} this can be a range.}

\item{comment}{A Comment object. See \code{\link[=createComment]{createComment()}}.}

\item{xy}{An alternative to specifying \code{col} and
\code{row} individually.  A vector of the form
\code{c(col, row)}.}

\item{gridExpand}{If \code{TRUE}, all data in rectangle min(rows):max(rows) X min(cols):max(cols)
will be removed.}
}
\description{
The comment functions (create, write and remove) allow the
modification of comments. In newer Excels they are called notes, while they
are called comments in openxml. Modification of what Excel now calls comment
(openxml calls them threadedComments) is not yet possible
}
\examples{
wb <- wb_workbook()
wb$addWorksheet("Sheet 1")

# write comment without author
c1 <- createComment(text = "this is a comment", author = "")
writeComment(wb, 1, col = "B", row = 10, comment = c1)

# Write another comment with author information
c2 <- createComment(text = "this is another comment", author = "Marco Polo")
writeComment(wb, 1, col = "C", row = 10, comment = c2)

# write a styled comment with system author
s1 <- create_font(b = "true", color = c(rgb = "FFFF0000"), sz = "12")
s2 <- create_font(color = c(rgb = "FF000000"), sz = "9")
c3 <- createComment(text = c("This Part Bold red\n\n", "This part black"), style = c(s1, s2))

writeComment(wb, 1, col = 6, row = 3, comment = c3)

# remove the first comment
removeComment(wb, 1, col = "B", row = 10)
\dontrun{
wb_save(wb, path = "createCommentExample.xlsx", overwrite = TRUE)
}
}
