% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{con_dims}
\alias{con_dims}
\title{create consecutive dims from column and row vector}
\usage{
con_dims(cols, rows)
}
\arguments{
\item{cols, rows}{integer vectors}
}
\description{
create consecutive dims from column and row vector
}
\keyword{internal}
