% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_border}
\alias{create_border}
\title{Helper to create a border}
\usage{
create_border(
  diagonal_down = "",
  diagonal_up = "",
  outline = "",
  bottom = NULL,
  bottom_color = NULL,
  diagonal = NULL,
  diagonal_color = NULL,
  end = "",
  horizontal = "",
  left = NULL,
  left_color = NULL,
  right = NULL,
  right_color = NULL,
  start = "",
  top = NULL,
  top_color = NULL,
  vertical = "",
  ...
)
}
\arguments{
\item{diagonal_down}{Logical, whether the diagonal border goes from the top left to the bottom right.}

\item{diagonal_up}{Logical, whether the diagonal border goes from the bottom left to the top right.}

\item{outline}{Logical, whether the border should be an outline border.}

\item{bottom}{Character, the style of the bottom border.}

\item{bottom_color}{Character, the color of the bottom border created with \code{\link[=wb_color]{wb_color()}}.}

\item{diagonal}{Character, the style of the diagonal border.}

\item{diagonal_color}{Character, the color of the diagonal border created with \code{\link[=wb_color]{wb_color()}}.}

\item{end}{Logical, whether the border is the end border.}

\item{horizontal}{Logical, whether the border is horizontal.}

\item{left}{Character, the style of the left border.}

\item{left_color}{Character, the color of the left border created with \code{\link[=wb_color]{wb_color()}}.}

\item{right}{Character, the style of the right border.}

\item{right_color}{Character, the color of the right border created with \code{\link[=wb_color]{wb_color()}}.}

\item{start}{Logical, whether the border is the start border.}

\item{top}{Character, the style of the top border.}

\item{top_color}{Character, the color of the top border created with \code{\link[=wb_color]{wb_color()}}.}

\item{vertical}{Logical, whether the border is vertical.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A formatted border object to be used in a spreadsheet.
}
\description{
This function creates border styles for a cell in a spreadsheet. Border styles can be any of the following: "none", "thin", "medium", "dashed", "dotted", "thick", "double", "hair", "mediumDashed", "dashDot", "mediumDashDot", "dashDotDot", "mediumDashDotDot", "slantDashDot". Border colors can be created with \code{\link[=wb_color]{wb_color()}}.
}
\examples{
# Create a border with a thick bottom and thin top
border <- create_border(
  bottom = "thick",
  bottom_color = wb_color("FF0000"),
  top = "thin",
  top_color = wb_color("00FF00")
)

}
\seealso{
\code{\link[=wb_add_border]{wb_add_border()}}

Other style creating functions: 
\code{\link{create_cell_style}()},
\code{\link{create_colors_xml}()},
\code{\link{create_dxfs_style}()},
\code{\link{create_fill}()},
\code{\link{create_font}()},
\code{\link{create_numfmt}()},
\code{\link{create_tablestyle}()}
}
\concept{style creating functions}
