% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{create_comment}
\alias{create_comment}
\title{Create a comment}
\usage{
create_comment(
  text,
  author = Sys.info()[["user"]],
  style = NULL,
  visible = TRUE,
  width = 2,
  height = 4
)
}
\arguments{
\item{text}{Comment text. Character vector.}

\item{author}{A string, by default, will use "user"}

\item{style}{A Style object or list of style objects the same length as comment vector.}

\item{visible}{Default: \code{TRUE}. Is the comment visible by default?}

\item{width}{Textbox integer width in number of cells}

\item{height}{Textbox integer height in number of cells}
}
\value{
a \code{wbComment} object
}
\description{
Use \code{\link[=wb_comment]{wb_comment()}} in new code.
}
\examples{
wb <- wb_workbook()
wb$add_worksheet("Sheet 1")

# write comment without author
c1 <- create_comment(text = "this is a comment", author = "")
wb$add_comment(dims = "B10", comment = c1)

# Write another comment with author information
c2 <- create_comment(text = "this is another comment", author = "Marco Polo")
wb$add_comment(sheet = 1, dims = "C10", comment = c2)

# write a styled comment with system author
s1 <- create_font(b = "true", color = wb_color(hex = "FFFF0000"), sz = "12")
s2 <- create_font(color = wb_color(hex = "FF000000"), sz = "9")
c3 <- create_comment(text = c("This Part Bold red\n\n", "This part black"), style = c(s1, s2))

wb$add_comment(sheet = 1, dims = wb_dims(3, 6), comment = c3)
}
\seealso{
\code{\link[=wb_add_comment]{wb_add_comment()}}
}
\keyword{internal}
