% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_numfmt}
\alias{create_numfmt}
\title{create number format}
\usage{
create_numfmt(numFmtId, formatCode)
}
\arguments{
\item{numFmtId}{an id}

\item{formatCode}{a format code}
}
\description{
create number format
}
