% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{delete_data}
\alias{delete_data}
\title{Defunct: Delete data}
\usage{
delete_data(wb, sheet, cols, rows)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet to clean}

\item{cols}{numeric column vector}

\item{rows}{numeric row vector}
}
\description{
This function is deprecated. Use \code{\link[=wb_clean_sheet]{wb_clean_sheet()}}.
}
\keyword{internal}
