% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{dims_to_dataframe}
\alias{dims_to_dataframe}
\title{Create dataframe from dimensions}
\usage{
dims_to_dataframe(dims, fill = FALSE, empty_rm = FALSE, cc = NULL)
}
\arguments{
\item{dims}{Character vector of expected dimension.}

\item{fill}{If \code{TRUE}, fills the dataframe with variables}

\item{empty_rm}{Logical if empty columns and rows should be included}

\item{cc}{A reference \code{cc} frame}
}
\description{
Non consecutive decreasing dims will return an increasing data frame.
}
\examples{
dims_to_dataframe("A1:B2")
}
\keyword{internal}
