% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt_txt}
\alias{fmt_txt}
\title{format strings independent of the cell style.}
\usage{
fmt_txt(
  x,
  bold = FALSE,
  italic = FALSE,
  underline = FALSE,
  strike = FALSE,
  size = NULL,
  color = NULL,
  font = NULL
)
}
\arguments{
\item{x}{a string or part of a string}

\item{bold}{bold}

\item{italic}{italic}

\item{underline}{underline}

\item{strike}{strike}

\item{size}{the font size}

\item{color}{a wbColor color for the font}

\item{font}{the font name}
}
\description{
format strings independent of the cell style.
}
\details{
The result is an xml string. It is possible to paste multiple \code{fmt_txt()}
strings together to create a string with multiple styles.
}
\examples{
fmt_txt("bla", underline = TRUE)
}
