% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{get_cell_refs}
\alias{get_cell_refs}
\title{Return excel cell coordinates from (x,y) coordinates}
\usage{
get_cell_refs(cellCoords)
}
\arguments{
\item{cellCoords}{A data.frame with two columns coordinate pairs.}
}
\value{
Excel alphanumeric cell reference
}
\description{
Return excel cell coordinates from (x,y) coordinates
}
\examples{
get_cell_refs(data.frame(1, 2))
# "B1"
get_cell_refs(data.frame(1:3, 2:4))
# "B1" "C2" "D3"
}
