% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{get_cell_style}
\alias{get_cell_style}
\title{helper get_cell_style}
\usage{
get_cell_style(wb, sheet, cell)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{a worksheet}

\item{cell}{a cell}
}
\description{
helper get_cell_style
}
