% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{get_date_origin}
\alias{get_date_origin}
\title{Get the date origin an xlsx file is using}
\usage{
get_date_origin(xlsxFile)
}
\arguments{
\item{xlsxFile}{An xlsx or xlsm file.}
}
\value{
One of "1900-01-01" or "1904-01-01".
}
\description{
Return the date origin used internally by an xlsx or xlsm file
}
\details{
Excel stores dates as the number of days from either 1904-01-01 or 1900-01-01. This function
checks the date origin being used in an Excel file and returns is so it can be used in \code{\link[=convert_date]{convert_date()}}
}
\examples{

## create a file with some dates
\dontrun{
write_xlsx(as.Date("2015-01-10") - (0:4), file = "get_date_originExample.xlsx")
m <- read_xlsx("get_date_originExample.xlsx")

## convert to dates
do <- get_date_origin(system.file("extdata", "readTest.xlsx", package = "openxlsx2"))
convert_date(m[[1]], do)
}
}
\seealso{
\code{\link[=convert_date]{convert_date()}}
}
