% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{get_styles}
\alias{get_styles}
\title{dummy function to create styleObjects}
\usage{
get_styles(styleObjects, wb)
}
\arguments{
\item{styleObjects}{styleObjects}

\item{wb}{wb}
}
\description{
dummy function to create styleObjects
}
