% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{getBaseFont}
\alias{getBaseFont}
\alias{wb_get_base_font}
\title{Return the workbook default font}
\usage{
wb_get_base_font(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\description{
Return the workbook default font

Returns the base font used in the workbook.
}
\examples{
## create a workbook
wb <- wb_workbook()
wb_get_base_font(wb)

## modify base font to size 10 Arial Narrow in red
modifyBaseFont(wb, fontSize = 10, fontColour = "#FF0000", fontName = "Arial Narrow")

wb_get_base_font(wb)
}
