% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{getStyles}
\alias{getStyles}
\title{Returns a list of all styles in the workbook}
\usage{
getStyles(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\description{
Returns list of style objects in the workbook
}
\examples{
## load a workbook
wb <- loadWorkbook(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))
getStyles(wb)[1:3]
}
\seealso{
\code{\link{replaceStyle}}
}
