% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{groupRows}
\alias{groupRows}
\title{Group Rows}
\usage{
groupRows(wb, sheet, rows, collapsed = FALSE)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{rows}{Indices of rows to group}

\item{collapsed}{Logical vector. If TRUE the grouped columns are collapsed. Defaults to FALSE}
}
\description{
Group a selection of rows
}
\examples{

# create matrix
t1 <- AirPassengers
t2 <- do.call(cbind, split(t1, cycle(t1)))
dimnames(t2) <- dimnames(.preformat.ts(t1))

wb <- createWorkbook()
addWorksheet(wb, "AirPass")
writeData(wb, "AirPass", t2)

# groups will always end on/show the last row. in the example 1950, 1955, and 1960
groupRows(wb, "AirPass", 2:3, collapsed = TRUE) # group years < 1950
groupRows(wb, "AirPass", 4:8, collapsed = TRUE) # group years 1951-1955
groupRows(wb, "AirPass", 9:13)                  # group years 1956-1960

}
\seealso{
\code{\link[=ungroupRows]{ungroupRows()}} to ungroup rows. \code{\link[=groupColumns]{groupColumns()}} for grouping columns.
}
