% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{init_named_style}
\alias{init_named_style}
\title{WIP init_named_style}
\usage{
init_named_style(wb, name)
}
\arguments{
\item{wb}{wbWorkbook}

\item{name}{style name}
}
\description{
WIP init_named_style
}
\details{
possible styles are:
"20\% - Accent1"
"20\% - Accent2"
"20\% - Accent3"
"20\% - Accent4"
"20\% - Accent5"
"20\% - Accent6"
"40\% - Accent1"
"40\% - Accent2"
"40\% - Accent3"
"40\% - Accent4"
"40\% - Accent5"
"40\% - Accent6"
"60\% - Accent1"
"60\% - Accent2"
"60\% - Accent3"
"60\% - Accent4"
"60\% - Accent5"
"60\% - Accent6"
"Accent1"
"Accent2"
"Accent3"
"Accent4"
"Accent5"
"Accent6"
"Bad"
"Calculation"
"Check Cell"
"Comma"
"Comma [0]"
"Currency"
"Currency [0]"
"Explanatory Text"
"Good"
"Heading 1"
"Heading 2"
"Heading 3"
"Heading 4"
"Input"
"Linked Cell"
”Neutral"
"Normal"
"Note"
"Output"
"Per cent"
"Title"
"Total"
"Warning Text"
}
