% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_dims}
\alias{is_dims}
\title{helper to check if a string looks like a cell}
\usage{
is_dims(x)
}
\arguments{
\item{x}{a string}
}
\description{
helper to check if a string looks like a cell
}
\keyword{internal}
