% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWorkbook.R
\name{loadWorkbook}
\alias{loadWorkbook}
\title{Load an existing .xlsx file}
\usage{
loadWorkbook(file, xlsxFile = NULL, isUnzipped = FALSE, sheet)
}
\arguments{
\item{file}{A path to an existing .xlsx or .xlsm file}

\item{xlsxFile}{alias for file}

\item{isUnzipped}{Set to TRUE if the xlsx file is already unzipped}

\item{sheet}{optional sheet parameter. if this is applied, only the selected
sheet will be loaded.}
}
\value{
Workbook object.
}
\description{
loadWorkbook returns a workbook object conserving styles and
formatting of the original .xlsx file.
}
\examples{
## load existing workbook from package folder
wb <- loadWorkbook(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))
names(wb) # list worksheets
wb ## view object
## Add a worksheet
wb_add_worksheet(wb, "A new worksheet")

## Save workbook
\dontrun{
wb_save(wb, "loadExample.xlsx", overwrite = TRUE)
}

}
\seealso{
\code{\link[=wb_remove_worksheet]{wb_remove_worksheet()}}
}
