% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{merge_cellXfs}
\alias{merge_cellXfs}
\title{merge cellXfs styles from workbook and new}
\usage{
merge_cellXfs(wb, new_cellxfs)
}
\arguments{
\item{wb}{a workbook}

\item{new_cellxfs}{new character vector of <xf ...>}
}
\description{
merge cellXfs styles from workbook and new
}
