% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names.wbWorkbook}
\alias{names.wbWorkbook}
\alias{names<-.wbWorkbook}
\title{get or set worksheet names
get or set worksheet names}
\usage{
\method{names}{wbWorkbook}(x)

\method{names}{wbWorkbook}(x) <- value
}
\arguments{
\item{x}{A \code{wbWorkbook} object}

\item{value}{a character vector the same length as \code{x}}
}
\description{
get or set worksheet names
get or set worksheet names
}
\examples{

wb <- wb_workbook()
wb$add_worksheet("S1")
wb$add_worksheet("S2")
wb$add_worksheet("S3")

names(wb)
names(wb)[[2]] <- "S2a"
names(wb)
names(wb) <- paste("Sheet", 1:3)
}
