% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{numfmt_is_date}
\alias{numfmt_is_date}
\title{check if numFmt is date. internal function}
\usage{
numfmt_is_date(numFmt)
}
\arguments{
\item{numFmt}{numFmt xml nodes}
}
\description{
check if numFmt is date. internal function
}
