% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{numfmt_is_posix}
\alias{numfmt_is_posix}
\title{check if numFmt is posix. internal function}
\usage{
numfmt_is_posix(numFmt)
}
\arguments{
\item{numFmt}{numFmt xml nodes}
}
\description{
check if numFmt is posix. internal function
}
