% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{op.openxlsx}
\alias{op.openxlsx}
\title{\code{openxlsx} options}
\format{
An object of class \code{list} of length 31.
}
\usage{
op.openxlsx
}
\description{
A named \code{list} of options that are set during \code{.onAttach()} and
are available for global controls of other functions.

\details{
\item{\code{bandedCols}}{Table style colun bands}
\item{\code{bandedRows}}{Table style row bands}
\item{\code{firstColumn}}{Table style, first column}
\item{\code{lastColumn}}{Table style, last column}
\item{\code{withFilter}}{Include filters on tables}
\item{\code{gridLines}}{Show grid lines}
\item{\code{tableStyle}}{Table style}

\item{\code{borders}}{Cell borders (e.g., \code{"none"}, \code{"sorrounding"}, \code{"all"})}
\item{\code{borderColour}}{Style border colour}
\item{\code{bordersStyle}}{Cell border style}

\item{\code{numFormat}}{Number format}
\item{\code{dateFormat}}{Date format style}
\item{\code{datetimeFormat}}{Date time format style}

\item{\code{header}}{Document headers; prioritized last}
\item{\code{footer}}{Document footers; prioritized last}
\item{\code{firstHeader}}{Document header; first page only; prioritized first}
\item{\code{firstFooter}}{Document footer; first page only; prioritized first}
\item{\code{evenFooter}}{Document footers; even pages}
\item{\code{evenHeader}}{Document headers; even pages}
\item{\code{oddFooter}}{Document footers; odd pages}
\item{\code{oddHeader}}{Document headers; odd pages}

\item{\code{dpi}}{Dots per inch (see \code{hdpi} and \code{vdpi} for other controls)}
\item{\code{hdpi}}{Horizontal dots per inch; prioritized over \code{dpi}}
\item{\code{vdpi}}{Vertical dots per inch; prioritized over \code{dpi}}

\item{\code{creator}}{Sets workbook creator (used in internal \code{get_creator()}; see \link{wbWorkbook} for more details)}
\item{\code{compressionLevel}}{Passed to \code{\link[zip:zip]{zip::zip()}}}
\item{\code{na}}{String to convert to \code{NA}}
\item{\code{keepNA}}{Convert \code{NA} to \code{"#N/A"} (or string in \code{na})}

\item{\code{orientation}}{paper orientation}
\item{\code{paperSize}}{paper size, in inches}
\item{\code{tabColour}}{tab colour}
}
}
\details{
A list of standard options from \code{openxlsx}
}
\keyword{datasets}
