% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2-package.R
\name{openxlsx2-defunct}
\alias{openxlsx2-defunct}
\title{Defunct functions in package \emph{openxlsx2}}
\description{
These functions were provided for compatibility with older versions of \code{openxlsx2},
and are now defunct. This guide helps you update your code to the latest standards.

As of openxlsx2 v1.0, API change should be minimal.
}
\section{Internal functions}{
These functions are used internally by openxlsx2. They originate from openxlsx, but
do not fit openxlsx2's API.

You should be able to modify
\itemize{
\item \code{\link[=delete_data]{delete_data()}} -> \code{\link[=wb_clean_sheet]{wb_clean_sheet()}}
\item \code{\link[=write_data]{write_data()}} -> \code{\link[=wb_add_data]{wb_add_data()}}
\item \code{\link[=write_datatable]{write_datatable()}} -> \code{\link[=wb_add_data_table]{wb_add_data_table()}}
\item \code{\link[=write_comment]{write_comment()}} -> \code{\link[=wb_add_comment]{wb_add_comment()}}
\item \code{\link[=remove_comment]{remove_comment()}} -> \code{\link[=wb_remove_comment]{wb_remove_comment()}}
\item \code{\link[=write_formula]{write_formula()}} -> \code{\link[=wb_add_formula]{wb_add_formula()}}
}

These functions were renamed for consistency.
\itemize{
\item \code{\link[=convertToExcelDate]{convertToExcelDate()}} -> \code{\link[=convert_to_excel_date]{convert_to_excel_date()}}
\item \code{\link[=wb_grid_lines]{wb_grid_lines()}} -> \code{\link[=wb_set_grid_lines]{wb_set_grid_lines()}}
\item \code{\link[=create_comment]{create_comment()}} -> \code{\link[=wb_comment]{wb_comment()}}
}
}

\seealso{
\link{.Defunct}
}
