% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2-package.R
\name{openxlsx2-deprecated}
\alias{openxlsx2-deprecated}
\title{Deprecated functions in package \emph{openxlsx2}}
\description{
While \code{openxlsx2} strives to keep its API semi-stable, as development
continues, there are inevitably changes. This guide helps you update your
code to the latest standards.
}
\section{Argument changes}{
For consistency, arguments were renamed to snake_case for the 0.8 release.
It is now recommended to use \code{dims} (the cell range) in favor of \code{row}, \code{col}, \code{start_row}, \code{start_col}

See \code{\link[=wb_dims]{wb_dims()}} as it provides many options on how to provide cell range
}

\section{Deprecated usage}{
\itemize{
\item \code{wb_get_named_regions()} will no longer allow providing a file.
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Before
wb_get_named_regions(file)

## Now
wb <- wb_load(file)
wb_get_named_regions(wb)
# also possible
wb_load(file)$get_named_regions()`
}\if{html}{\out{</div>}}
}

\seealso{
\link{.Deprecated}
}
