% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2.R
\docType{package}
\name{openxlsx2}
\alias{openxlsx2}
\title{xlsx reading, writing and editing.}
\description{
openxlsx simplifies the the process of writing and styling Excel xlsx files from R
and removes the dependency on Java.
}
\details{
The openxlsx package uses global options to simplify formatting:

\itemize{
\item{\code{options("openxlsx2.borderColor" = "black")}}
\item{\code{options("openxlsx2.borderStyle" = "thin")}}
\item{\code{options("openxlsx2.dateFormat" = "mm/dd/yyyy")}}
\item{\code{options("openxlsx2.datetimeFormat" = "yyyy-mm-dd hh:mm:ss")}}
\item{\code{options("openxlsx2.numFmt" = NULL)}}
\item{\code{options("openxlsx2.paperSize" = 9)}} ## A4
\item{\code{options("openxlsx2.orientation" = "portrait")}} ## page orientation
}
See the Formatting vignette for examples.

Additional options

\itemize{
\item{\code{options("openxlsx2.compressionLevel" = "9")}} ## set zip compression level, default is "1".
}
}
\seealso{
\itemize{
\item{\code{vignette("Introduction", package = "openxlsx2")}}
\item{\code{vignette("formatting", package = "openxlsx2")}}
\item{\code{\link[=write_data]{write_data()}}}
\item{\code{\link[=write_datatable]{write_datatable()}}}
\item{\code{\link[=write_xlsx]{write_xlsx()}}}
\item{\code{\link[=read_xlsx]{read_xlsx()}}}
}
for examples
}
