% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2-package.R
\name{openxlsx2_options}
\alias{openxlsx2_options}
\title{openxlsx2 Options}
\description{
The openxlsx2 package uses global options to simplify formatting:
}
\details{
\itemize{
\item \code{options("openxlsx2.borderColor" = "black")}
\item \code{options("openxlsx2.borderStyle" = "thin")}
\item \code{options("openxlsx2.dateFormat" = "mm/dd/yyyy")}
\item \code{options("openxlsx2.datetimeFormat" = "yyyy-mm-dd hh:mm:ss")}
\item \code{options("openxlsx2.numFmt" = NULL)}
\item \code{options("openxlsx2.paperSize" = 9)} ## A4
\item \code{options("openxlsx2.orientation" = "portrait")} ## page orientation
\item \code{options("openxlsx2.sheet.default_name" = "Sheet")}
\item \code{options("openxlsx2.rightToLeft" = NULL)}
}
}
