% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{person-wb}
\alias{person-wb}
\alias{wb_add_person}
\alias{wb_get_person}
\title{Helper for adding threaded comments}
\usage{
wb_add_person(wb, name = NULL, id = NULL, user_id = NULL, provider_id = "None")

wb_get_person(wb, name = NULL)
}
\arguments{
\item{wb}{a Workbook}

\item{name}{the name of the person to display.}

\item{id}{(optional) the display id}

\item{user_id}{(optional) the user id}

\item{provider_id}{(optional) the provider id}
}
\description{
Adds a person to a workbook, so that they can be the author of threaded
comments in a workbook with \code{\link[=wb_add_thread]{wb_add_thread()}}
}
\seealso{
\code{\link[=wb_add_thread]{wb_add_thread()}}
}
\keyword{comments}
