% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{print.pugi_xml}
\alias{print.pugi_xml}
\title{print pugi_xml}
\usage{
\method{print}{pugi_xml}(x, indent = " ", raw = FALSE, ...)
}
\arguments{
\item{x}{something to print}

\item{indent}{indent used default is " "}

\item{raw}{print as raw text}

\item{...}{to please check}
}
\description{
print pugi_xml
}
\examples{
  # a pointer
  x <- read_xml("<a><b/></a>")
  print(x)
  print(x, raw = TRUE)
}
