% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{properties}
\alias{properties}
\alias{wb_get_properties}
\alias{wb_set_properties}
\title{Modify workbook properties}
\usage{
wb_get_properties(wb)

wb_set_properties(
  wb,
  creators = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  date_time_created = Sys.time(),
  modifiers = NULL,
  keywords = NULL,
  comments = NULL
)
}
\arguments{
\item{wb}{A Workbook object}

\item{creators}{A character string indicating who has created the workbook}

\item{title, subject, category, keywords, comments}{Workbook property, a string.}

\item{date_time_created}{datetime created}

\item{modifiers}{A character string indicating who was the last person to modify the workbook}
}
\value{
A wbWorkbook object, invisibly.
}
\description{
This function is useful for workbooks that are loaded. It can be used to set the
workbook \code{title}, \code{subject} and \code{category} field. Use \code{\link[=wb_workbook]{wb_workbook()}}
to easily set these properties with a new workbook.
}
\examples{
file <- system.file("extdata", "openxlsx2_example.xlsx", package = "openxlsx2")
wb <- wb_load(file)
wb$get_properties()

# Add a title to properties
wb$set_properties(title = "my title")
wb$get_properties()
}
\seealso{
\code{\link[=wb_workbook]{wb_workbook()}}
}
