% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{removeColWidths}
\alias{removeColWidths}
\title{Remove column widths from a worksheet}
\usage{
removeColWidths(wb, sheet, cols)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Indices of columns to remove custom width (if any) from.}
}
\description{
Remove column widths from a worksheet
}
\examples{
## Create a new workbook
wb <- wb_load(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))

## remove column widths in columns 1 to 20
removeColWidths(wb, 1, cols = 1:20)
\dontrun{
wb_save(wb, "removeColWidthsExample.xlsx", overwrite = TRUE)
}
}
\seealso{
\code{\link[=setColWidths]{setColWidths()}}
}
