% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{removeFilter}
\alias{removeFilter}
\title{Remove a worksheet filter}
\usage{
removeFilter(wb, sheet)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A vector of names or indices of worksheets}
}
\description{
Removes filters from addFilter() and writeData()
}
\examples{
wb <- wb_workbook()
wb$addWorksheet("Sheet 1")
wb$addWorksheet("Sheet 2")
wb$addWorksheet("Sheet 3")

writeData(wb, 1, iris)
addFilter(wb, 1, row = 1, cols = seq_along(iris))

## Equivalently
writeData(wb, 2, x = iris, withFilter = TRUE)

## Similarly
writeDataTable(wb, 3, iris)

## remove filters
removeFilter(wb, 1:2) ## remove filters
removeFilter(wb, 3) ## Does not affect tables!
\dontrun{
wb_save(wb, path = "removeFilterExample.xlsx", overwrite = TRUE)
}
}
