% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{setLastModifiedBy}
\alias{setLastModifiedBy}
\title{Add another author to the meta data of the file.}
\usage{
setLastModifiedBy(wb, LastModifiedBy)
}
\arguments{
\item{wb}{A workbook object}

\item{LastModifiedBy}{A string object with the name of the LastModifiedBy-User}
}
\description{
Just a wrapper of wb$changeLastModifiedBy()
}
\examples{
wb <- wb_workbook()
setLastModifiedBy(wb, "test")
}
\seealso{
Other workbook wrappers: 
\code{\link{freezePane}},
\code{\link{setColWidths}()},
\code{\link{wb_add_worksheet}()},
\code{\link{wb_clone_worksheet}()},
\code{\link{wb_creators}},
\code{\link{wb_get_base_font}()},
\code{\link{wb_save}()},
\code{\link{wb_set_row_heights}()},
\code{\link{wb_workbook}()},
\code{\link{workbook_grouping}},
\code{\link{ws_cell_merge}}
}
\concept{workbook wrappers}
