% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{style_is_posix}
\alias{style_is_posix}
\title{check if style is posix. internal function}
\usage{
style_is_posix(cellXfs, numfmt_date)
}
\arguments{
\item{cellXfs}{cellXfs xml nodes}

\item{numfmt_date}{custom numFmtId dates}
}
\description{
check if style is posix. internal function
}
