% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{styles_on_sheet}
\alias{styles_on_sheet}
\title{get all styles on a sheet}
\usage{
styles_on_sheet(wb, sheet)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{worksheet}
}
\description{
get all styles on a sheet
}
