% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-utils.R
\name{waivers}
\alias{waivers}
\alias{current_sheet}
\alias{next_sheet}
\alias{na_strings}
\title{\code{openxlsx2} waivers}
\usage{
current_sheet()

next_sheet()

na_strings()
}
\value{
An object of class \code{openxlsx2_waiver}
}
\description{
Waiver functions for \code{openxlsx2} functions.
\itemize{
\item \code{current_sheet()} uses \code{\link[=wb_get_active_sheet]{wb_get_active_sheet()}} by default if performing
actions on a worksheet, for example when you add data.
\item \code{next_sheet()} is used when you add a new worksheet, a new chartsheet or
when you add a pivot table
}
}
