% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_add_chart}
\alias{wb_add_chart}
\title{dummy function to add a chart to an existing workbook
currently only a barplot is possible}
\usage{
wb_add_chart(wb, sheet, srhc, dims, type = "barplot")
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{the sheet on which the graph will appear}

\item{srhc}{a somewhat complex created list object containing data}

\item{dims}{the dimensions where the sheet will appear}

\item{type}{the chart type, currently only "barplot"}
}
\description{
dummy function to add a chart to an existing workbook
currently only a barplot is possible
}
\examples{
wb <- wb_workbook()$
  add_worksheet("S1")$add_data("S1", cars)$
  add_worksheet("S2")$add_data("S2", mtcars, rowNames = TRUE)

# create Chart input
srhc <- list(
  list(sheet = "'S2'",
       head = "$B$1",
       body = "$B$2:$B$33",
       cat = "$A$2:$A$33")
  ,list(sheet = "'S2'",
        head = "$C$1",
        body = "$C$2:$C33",
        cat = "$A$2:$A$33")
)

wb <- wb_add_chart(wb, sheet = 2, srhc, dims = "A1:H18")

}
