% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_comment}
\alias{wb_add_comment}
\title{Add comment to worksheet}
\usage{
wb_add_comment(wb, sheet = current_sheet(), dims = "A1", comment, ...)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A worksheet of the workbook}

\item{dims}{Optional row and column as spreadsheet dimension, e.g. "A1"}

\item{comment}{A comment to apply to the worksheet}

\item{...}{additional arguments}
}
\value{
The \code{wbWorkbook} object
}
\description{
Add comment to worksheet
}
\seealso{
\code{\link[=wb_add_thread]{wb_add_thread()}}
}
