% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_drawing}
\alias{wb_add_drawing}
\title{add drawings to workbook}
\usage{
wb_add_drawing(wb, sheet = current_sheet(), xml, dims = "A1:H8")
}
\arguments{
\item{wb}{a wbWorkbook}

\item{sheet}{a sheet in the workbook}

\item{xml}{the drawing xml as character or file}

\item{dims}{the dimension where the drawing is added. Can be NULL}
}
\description{
add drawings to workbook
}
