% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_header_footer_image}
\alias{wb_add_header_footer_image}
\title{Add an image to a worksheet header or footer}
\usage{
wb_add_header_footer_image(
  wb,
  sheet = current_sheet(),
  file,
  position = c("left", "center", "right"),
  location = c("header", "footer"),
  width = 2,
  height = 1,
  units = c("in", "cm", "pt")
)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A worksheet of the workbook}

\item{file}{An image file in png, jpeg, gif, bmp, emf, or wmf.}

\item{position}{The section the image is placed in, left, center or right}

\item{location}{Either header or footer}

\item{width, height}{The size of the image. Excel stores a size, not a scale
factor, so a picture is stretched to whatever is given here.}

\item{units}{The units of \code{width} and \code{height}, in, cm or pt}
}
\value{
The \code{wbWorkbook} object, invisibly
}
\description{
Places an image in one of the three sections of a worksheet header or
footer. Header images are shown in the page layout view and in print output.
}
\examples{
img <- system.file("extdata", "einstein.jpg", package = "openxlsx2")

wb <- wb_workbook()$add_worksheet()
wb <- wb_add_header_footer_image(wb, file = img, position = "right")
wb <- wb_set_sheetview(wb, view = "pageLayout")
}
\seealso{
\code{\link[=wb_add_image]{wb_add_image()}} \code{\link[=wb_set_header_footer]{wb_set_header_footer()}}

Other workbook wrappers:
\code{\link{base_font-wb}},
\code{\link{col_widths-wb}},
\code{\link{creators-wb}},
\code{\link{grouping-wb}},
\code{\link{row_heights-wb}},
\code{\link[=wb_add_chartsheet]{wb_add_chartsheet()}},
\code{\link[=wb_add_data]{wb_add_data()}},
\code{\link[=wb_add_data_table]{wb_add_data_table()}},
\code{\link[=wb_add_formula]{wb_add_formula()}},
\code{\link[=wb_add_hyperlink]{wb_add_hyperlink()}},
\code{\link[=wb_add_pivot_table]{wb_add_pivot_table()}},
\code{\link[=wb_add_slicer]{wb_add_slicer()}},
\code{\link[=wb_add_worksheet]{wb_add_worksheet()}},
\code{\link{wb_base_colors}},
\code{\link[=wb_clone_worksheet]{wb_clone_worksheet()}},
\code{\link[=wb_copy_cells]{wb_copy_cells()}},
\code{\link[=wb_freeze_pane]{wb_freeze_pane()}},
\code{\link[=wb_merge_cells]{wb_merge_cells()}},
\code{\link[=wb_save]{wb_save()}},
\code{\link[=wb_set_last_modified_by]{wb_set_last_modified_by()}},
\code{\link[=wb_workbook]{wb_workbook()}}
}
\concept{workbook wrappers}
