% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{wb_add_named_style}
\alias{wb_add_named_style}
\title{WIP wb_add_named_style}
\usage{
wb_add_named_style(wb, sheet = current_sheet(), dims = "A1", name)
}
\arguments{
\item{wb}{wbWorkbook}

\item{sheet}{a worksheet}

\item{dims}{worksheet dimensions}

\item{name}{a style name}
}
\description{
WIP wb_add_named_style
}
