% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{wb_add_person}
\alias{wb_add_person}
\title{Add person to use for threaded comment}
\usage{
wb_add_person(wb, name = NULL, id = NULL, user_id = NULL, provider_id = "None")
}
\arguments{
\item{wb}{a workbook}

\item{name}{the name to display}

\item{id}{(optional) the display id}

\item{user_id}{(optional) the user id}

\item{provider_id}{(optional) the provider id}
}
\description{
If a threaded comment is added, it needs a person attached with it. The default is to create a person with provider id \code{"None"}. Other providers are possible with specific values for \code{id} and \code{user_id}. If you require the following, create a workbook via spreadsheet software load it and get the values with \code{wb_get_person()}
}
\keyword{comments}
