% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_slicer}
\alias{wb_add_slicer}
\alias{wb_remove_slicer}
\alias{wb_add_timeline}
\alias{wb_remove_timeline}
\title{Add a slicer/timeline to a pivot table}
\usage{
wb_add_slicer(
  wb,
  x,
  dims = "A1",
  sheet = current_sheet(),
  pivot_table,
  slicer,
  params
)

wb_remove_slicer(wb, sheet = current_sheet())

wb_add_timeline(
  wb,
  x,
  dims = "A1",
  sheet = current_sheet(),
  pivot_table,
  timeline,
  params
)

wb_remove_timeline(wb, sheet = current_sheet())
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{A \code{data.frame} that inherits the \code{\link{wb_data}} class.}

\item{dims}{The worksheet cell where the pivot table is placed}

\item{sheet}{A worksheet}

\item{pivot_table}{The name of a pivot table}

\item{slicer, timeline}{A variable used as slicer/timeline for the pivot table}

\item{params}{A list of parameters to modify pivot table creation. See \strong{Details} for available options.}
}
\description{
Add a slicer/timeline to a previously created pivot table. This function is still experimental and might be changed/improved in upcoming releases.
}
\details{
This assumes that the slicer/timeline variable initialization has happened before. Unfortunately, it is unlikely that we can guarantee this for loaded workbooks, and we \emph{strictly} discourage users from attempting this. If the variable has not been initialized properly, this may cause the spreadsheet software to crash.
Although it is documented that slicers should use "TimelineStyleLight[1-6]" and "TimelineStyleDark[1-6]" they use slicer styles.

Possible \code{params} arguments for slicers are listed below.
\itemize{
\item edit_as: "twoCell" to place the slicer into the cells
\item column_count: integer used as column count
\item sort_order: "descending" / "ascending"
\item choose: select variables in the form of a named logical vector like
\code{c(agegp = 'x > "25-34"')} for the \code{esoph} dataset.
\item locked_position
\item start_item
\item hide_no_data_items
}

Possible \code{params} arguments for timelines are listed below.
\itemize{
\item beg_date/end_date: dates when the timeline should begin or end
\item choose_beg/choose_end: dates when the selection should begin or end
\item scroll_position
\item show_selection_label
\item show_time_level
\item show_horizontal_scrollbar
}

Possible common \code{params}:
\itemize{
\item caption: string used for a caption
\item style: "SlicerStyleLight[1-6]", "SlicerStyleDark[1-6]" only for slicer "SlicerStyleOther[1-2]"
\item level: the granularity of the slicer (for timeline 0 = year, 1 = quarter, 2 = month)
\item show_caption: logical if caption should be shown or not
}

Removing works on the spreadsheet level. Therefore all slicers/timelines are removed from a worksheet. At the moment the drawing reference remains on the spreadsheet. Therefore spreadsheet software that does not handle slicers/timelines will still show the drawing.
}
\examples{
# prepare data
df <- data.frame(
  AirPassengers = c(AirPassengers),
  time = seq(from = as.Date("1949-01-01"), to = as.Date("1960-12-01"), by = "month"),
  letters = letters[1:4],
  stringsAsFactors = FALSE
)

# create workbook
wb <- wb_workbook()$
  add_worksheet("pivot")$
  add_worksheet("data")$
  add_data(x = df)

# get pivot table data source
df <- wb_data(wb, sheet = "data")

# create pivot table
wb$add_pivot_table(
  df,
  sheet = "pivot",
  rows = "time",
  cols = "letters",
  data = "AirPassengers",
  pivot_table = "airpassengers",
  params = list(
    compact = FALSE, outline = FALSE, compact_data = FALSE,
    row_grand_totals = FALSE, col_grand_totals = FALSE)
)

# add slicer
wb$add_slicer(
  df,
  dims = "E1:I7",
  sheet = "pivot",
  slicer = "letters",
  pivot_table = "airpassengers",
  params = list(choose = c(letters = 'x \%in\% c("a", "b")'))
)

# add timeline
wb$add_timeline(
  df,
  dims = "E9:I14",
  sheet = "pivot",
  timeline = "time",
  pivot_table = "airpassengers",
  params = list(
    beg_date = as.Date("1954-01-01"),
    end_date = as.Date("1961-01-01"),
    choose_beg = as.Date("1957-01-01"),
    choose_end = as.Date("1958-01-01"),
    level = 0,
    style = "TimeSlicerStyleLight2"
  )
)
}
\seealso{
Other workbook wrappers:
\code{\link{base_font-wb}},
\code{\link{col_widths-wb}},
\code{\link{creators-wb}},
\code{\link{grouping-wb}},
\code{\link{row_heights-wb}},
\code{\link[=wb_add_chartsheet]{wb_add_chartsheet()}},
\code{\link[=wb_add_data]{wb_add_data()}},
\code{\link[=wb_add_data_table]{wb_add_data_table()}},
\code{\link[=wb_add_formula]{wb_add_formula()}},
\code{\link[=wb_add_header_footer_image]{wb_add_header_footer_image()}},
\code{\link[=wb_add_hyperlink]{wb_add_hyperlink()}},
\code{\link[=wb_add_pivot_table]{wb_add_pivot_table()}},
\code{\link[=wb_add_worksheet]{wb_add_worksheet()}},
\code{\link{wb_base_colors}},
\code{\link[=wb_clone_worksheet]{wb_clone_worksheet()}},
\code{\link[=wb_copy_cells]{wb_copy_cells()}},
\code{\link[=wb_freeze_pane]{wb_freeze_pane()}},
\code{\link[=wb_merge_cells]{wb_merge_cells()}},
\code{\link[=wb_save]{wb_save()}},
\code{\link[=wb_set_last_modified_by]{wb_set_last_modified_by()}},
\code{\link[=wb_workbook]{wb_workbook()}}

Other worksheet content functions:
\code{\link{col_widths-wb}},
\code{\link{filter-wb}},
\code{\link{grouping-wb}},
\code{\link{named_region-wb}},
\code{\link{row_heights-wb}},
\code{\link[=wb_add_conditional_formatting]{wb_add_conditional_formatting()}},
\code{\link[=wb_add_data]{wb_add_data()}},
\code{\link[=wb_add_data_table]{wb_add_data_table()}},
\code{\link[=wb_add_formula]{wb_add_formula()}},
\code{\link[=wb_add_hyperlink]{wb_add_hyperlink()}},
\code{\link[=wb_add_pivot_table]{wb_add_pivot_table()}},
\code{\link[=wb_add_thread]{wb_add_thread()}},
\code{\link[=wb_freeze_pane]{wb_freeze_pane()}},
\code{\link[=wb_merge_cells]{wb_merge_cells()}}
}
\concept{workbook wrappers}
\concept{worksheet content functions}
