% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_style_across}
\alias{wb_add_style_across}
\title{Set the column / row style in a worksheet}
\usage{
wb_add_style_across(
  wb,
  sheet = current_sheet(),
  from = "A1",
  cols = NULL,
  rows = NULL
)
}
\arguments{
\item{wb}{A workbook}

\item{sheet}{A worksheet}

\item{from}{The cell containing the style that should be applied column-wise/row-wise}

\item{cols}{The columns the style will be applied to, either "A:D" or 1:4}

\item{rows}{The rows the style will be applied to}
}
\description{
Apply the style from one cell across columns/rows of the worksheet. This affects only empty cells.
}
\examples{
wb <- wb_workbook() \%>\%
  wb_add_worksheet() \%>\%
  wb_add_fill(dims = "C3", color = wb_color("yellow")) \%>\%
  wb_add_style_across(from = "C3", cols = "C:D", rows = 3:4)
}
\seealso{
Other workbook styling functions: 
\code{\link{base_font-wb}},
\code{\link{wb_add_dxfs_style}()},
\code{\link{wb_add_style}()}
}
\concept{workbook styling functions}
