% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{wb_add_threaded_comment}
\alias{wb_add_threaded_comment}
\title{add threaded comment to worksheet}
\usage{
wb_add_threaded_comment(
  wb,
  sheet = current_sheet(),
  dims = "A1",
  comment = NULL,
  person_id,
  reply = FALSE,
  resolve = FALSE
)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{a worksheet}

\item{dims}{a cell}

\item{comment}{the comment to add}

\item{person_id}{the person Id this should be added for}

\item{reply}{logical if the comment is a reply}

\item{resolve}{logical if the comment should be maked as resolved}
}
\description{
add threaded comment to worksheet
}
