% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{wb_comment}
\alias{wb_comment}
\title{Create a comment object}
\usage{
wb_comment(
  text = NULL,
  style = NULL,
  visible = TRUE,
  author = getOption("openxlsx2.creator"),
  width = 2,
  height = 4
)
}
\arguments{
\item{text}{Comment text. Character vector.}

\item{style}{A Style object or list of style objects the same length as comment vector.}

\item{visible}{\code{TRUE} or \code{FALSE}. Is comment visible?}

\item{author}{Author of comment. A string. By default, will look at \code{options("openxlsx2.creator")}.
Otherwise, will check the system username.}

\item{width}{Textbox integer width in number of cells}

\item{height}{Textbox integer height in number of cells}
}
\value{
A \code{wbComment} object
}
\description{
Create a comment object
}
