% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_creators}
\alias{wb_creators}
\alias{wb_add_creator}
\alias{wb_set_creator}
\alias{wb_remove_creator}
\alias{wb_get_creator}
\title{Workbook creators}
\usage{
wb_add_creator(wb, creator)

wb_set_creator(wb, creator)

wb_remove_creator(wb, creator)

wb_get_creator(wb)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object}

\item{creator}{A character vector of names}
}
\value{
\itemize{
\item \code{wb_set_creator()}, \code{wb_add_creator()}, and \code{wb_remove_creator()} return the \code{wbWorkbook} object
\item \code{wb_get_creator()} returns a \code{character} vector of creators
}
}
\description{
Modify and get workbook creators
}
\examples{

# workbook made with default creator (see [wbWorkbook])
wb <- wb_workbook()
wb_get_creator(wb)

# add a new creator (assuming "test" isn't your default creator)
wb <- wb_add_creator(wb, "test")
wb_get_creator(wb)

# remove the creator
wb <- wb_remove_creator(wb, "test")
wb_get_creator(wb)

}
\concept{workbook_wrappers}
