% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_data}
\alias{wb_data}
\title{provide wb_data object for mschart}
\usage{
wb_data(wb, sheet = current_sheet(), dims = "A1")
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{a sheet in the workbook either name or index}

\item{dims}{the dimensions}
}
\description{
provide wb_data object for mschart
}
