% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wb_dims}
\alias{wb_dims}
\title{Helper to specify the \code{dims} argument.}
\usage{
wb_dims(...)
}
\arguments{
\item{...}{construct dims arguments, from rows/cols vectors or objects that
can be coerced to data frame}
}
\value{
A \code{dims} string
}
\description{
\code{wb_dims()} can be used to help provide the \code{dims} argument, in the \verb{wb_add_*} functions.
It returns a Excel range (i.e. "A1:B1") or a start like "A2".
It can be very useful as you can specify many parameters that interact together
In general, you must provide named arguments. \code{wb_dims()} will only accept unnamed arguments
if it is \code{rows}, \code{cols}, for example \code{wb_dims(1:4, 1:2)}, that will return "A1:B4".

\code{wb_dims()} can also be used with an object (a \code{data.frame} or a \code{matrix} for example.)
All parameters are numeric unless stated otherwise.
}
\section{Using \code{wb_dims()} alone}{
\itemize{
\item \code{rows}
\item \code{cols}
\item \code{start_row}
\item \code{start_col}
}
}

\section{Using \code{wb_dims()} with an object}{
\itemize{
\item \code{x}
\item \code{start_row} the starting row of \code{x} (The \code{dims} returned will be )
\item \code{start_col} the starting column: a single integer, or an Excel column identifier "A", "B" etc.
\item \code{rows} (optional)
\item \code{cols}
\item \code{row_names}  A logical, should include \code{row_names}
\item \code{col_names} A logical, should include \code{col_names} (can be useful to style only the data.)
}
}

\examples{
# Provide  vectors
wb_dims(1:10, 1:5)
wb_dims(rows = 1:10, cols = 1:10)
# provide `start_col`
wb_dims(rows = 1:10, cols = 1:10, start_row = 2)
# or objects
wb_dims(x = mtcars)
wb_dims()
}
