% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_get_cell_style}
\alias{wb_get_cell_style}
\title{get cell style}
\usage{
wb_get_cell_style(wb, sheet, dims)
}
\arguments{
\item{wb}{wb}

\item{sheet}{sheet}

\item{dims}{dims}
}
\description{
get cell style
}
