% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{wb_get_mips}
\alias{wb_get_mips}
\title{function to extract mips string from workbook}
\usage{
wb_get_mips(wb, single_xml = TRUE)
}
\arguments{
\item{wb}{a workbook}

\item{single_xml}{option to define if the string should be exported as single string. helpful if storing as option is desired.}
}
\description{
function to extract mips string from workbook
}
\seealso{
wb_add_mips
}
