% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{wb_get_person}
\alias{wb_get_person}
\title{Get Person list from workbook}
\usage{
wb_get_person(wb, name = NULL)
}
\arguments{
\item{wb}{a workbook}

\item{name}{a specific name}
}
\description{
Persons are required for threaded comments
}
