% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook.R
\name{wb_get_sheet_name}
\alias{wb_get_sheet_name}
\title{Get sheet name}
\usage{
wb_get_sheet_name(wb, index = NULL)
}
\arguments{
\item{wb}{a \link{wbWorkbook} object}

\item{index}{Sheet name index}
}
\value{
The sheet index
}
\description{
Get sheet name
}
