% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_get_sheet_names}
\alias{wb_get_sheet_names}
\title{Get worksheet names for a workbook}
\usage{
wb_get_sheet_names(wb)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}
}
\value{
A \code{named} \code{character} vector of sheet names in their order.  The
names represent the original value of the worksheet prior to any character
substitutions.
}
\description{
Gets the worksheet names for a \link{wbWorkbook} object
}
