% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_get_worksheet}
\alias{wb_get_worksheet}
\alias{wb_ws}
\title{Get a worksheet from a \code{wbWorkbook} object}
\usage{
wb_get_worksheet(wb, sheet)

wb_ws(wb, sheet)
}
\arguments{
\item{wb}{a \link{wbWorkbook} object}

\item{sheet}{A sheet name or index}
}
\value{
A \code{wbWorksheet} object
}
\description{
Get a worksheet from a \code{wbWorkbook} object
}
