% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_grid_lines}
\alias{wb_grid_lines}
\title{Set worksheet gridlines to show or hide.}
\usage{
wb_grid_lines(wb, sheet = current_sheet(), show = FALSE)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{show}{A logical. If \code{FALSE}, grid lines are hidden.}
}
\description{
Set worksheet gridlines to show or hide.
}
\examples{
wb <- wb_load(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))
wb$get_sheet_names() ## list worksheets in workbook
wb$grid_lines(1, show = FALSE)
wb$grid_lines("testing", show = FALSE)
}
