% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_merge_cells}
\alias{wb_merge_cells}
\alias{wb_unmerge_cells}
\title{Merge cells within a worksheet}
\usage{
wb_merge_cells(
  wb,
  sheet = current_sheet(),
  dims = NULL,
  solve = FALSE,
  direction = NULL,
  ...
)

wb_unmerge_cells(wb, sheet = current_sheet(), dims = NULL, ...)
}
\arguments{
\item{wb}{A Workbook object}

\item{sheet}{A name or index of a worksheet}

\item{dims}{worksheet cells}

\item{solve}{logical if intersecting merges should be solved}

\item{direction}{direction in which to split the cell merging. Allows "row" or "col"}

\item{...}{additional arguments}
}
\description{
Worksheet cell merging
}
\details{
If using the deprecated arguments \code{rows} and \code{cols} with a merged region must be rectangular,
only min and max of \code{cols} and \code{rows} are used.
}
\examples{
# Create a new workbook
wb <- wb_workbook()$add_worksheet()

# Merge cells: Row 2 column C to F (3:6)
wb <- wb_merge_cells(wb, dims = "C3:F6")

# Merge cells:Rows 10 to 20 columns A to J (1:10)
wb <- wb_merge_cells(wb, dims = wb_dims(rows = 10:20, cols = 1:10))

wb$add_worksheet()

## Intersecting merges
wb <- wb_merge_cells(wb, dims = wb_dims(cols = 1:10, rows = 1))
wb <- wb_merge_cells(wb, dims = wb_dims(cols = 5:10, rows = 2))
wb <- wb_merge_cells(wb, dims = wb_dims(cols = 1:10, rows = 12))
try(wb_merge_cells(wb, dims = "A1:A10"))

## remove merged cells
# removes any intersecting merges
wb <- wb_unmerge_cells(wb, dims = wb_dims(cols = 1, rows = 1))
wb <- wb_merge_cells(wb, dims = "A1:A10")

# or let us decide how to solve this
wb <- wb_merge_cells(wb, dims = "A1:A10", solve = TRUE)

}
\seealso{
Other workbook wrappers:
\code{\link{base_font-wb}},
\code{\link{col_widths-wb}},
\code{\link{creators-wb}},
\code{\link{grouping-wb}},
\code{\link{row_heights-wb}},
\code{\link[=wb_add_chartsheet]{wb_add_chartsheet()}},
\code{\link[=wb_add_data]{wb_add_data()}},
\code{\link[=wb_add_data_table]{wb_add_data_table()}},
\code{\link[=wb_add_formula]{wb_add_formula()}},
\code{\link[=wb_add_header_footer_image]{wb_add_header_footer_image()}},
\code{\link[=wb_add_hyperlink]{wb_add_hyperlink()}},
\code{\link[=wb_add_pivot_table]{wb_add_pivot_table()}},
\code{\link[=wb_add_slicer]{wb_add_slicer()}},
\code{\link[=wb_add_worksheet]{wb_add_worksheet()}},
\code{\link{wb_base_colors}},
\code{\link[=wb_clone_worksheet]{wb_clone_worksheet()}},
\code{\link[=wb_copy_cells]{wb_copy_cells()}},
\code{\link[=wb_freeze_pane]{wb_freeze_pane()}},
\code{\link[=wb_save]{wb_save()}},
\code{\link[=wb_set_last_modified_by]{wb_set_last_modified_by()}},
\code{\link[=wb_workbook]{wb_workbook()}}

Other worksheet content functions:
\code{\link{col_widths-wb}},
\code{\link{filter-wb}},
\code{\link{grouping-wb}},
\code{\link{named_region-wb}},
\code{\link{row_heights-wb}},
\code{\link[=wb_add_conditional_formatting]{wb_add_conditional_formatting()}},
\code{\link[=wb_add_data]{wb_add_data()}},
\code{\link[=wb_add_data_table]{wb_add_data_table()}},
\code{\link[=wb_add_formula]{wb_add_formula()}},
\code{\link[=wb_add_hyperlink]{wb_add_hyperlink()}},
\code{\link[=wb_add_pivot_table]{wb_add_pivot_table()}},
\code{\link[=wb_add_slicer]{wb_add_slicer()}},
\code{\link[=wb_add_thread]{wb_add_thread()}},
\code{\link[=wb_freeze_pane]{wb_freeze_pane()}}
}
\concept{workbook wrappers}
\concept{worksheet content functions}
