% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_page_break}
\alias{wb_page_break}
\title{add a page break to a worksheet}
\usage{
wb_page_break(wb, sheet, i, type = "row")
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{i}{row or column number to insert page break.}

\item{type}{One of "row" or "column" for a row break or column break.}
}
\description{
insert page breaks into a worksheet
}
\examples{
wb <- wb_workbook()
wb$add_worksheet("Sheet 1")
write_data(wb, sheet = 1, x = iris)

wb_page_break(wb, sheet = 1, i = 10, type = "row")
wb_page_break(wb, sheet = 1, i = 20, type = "row")
wb_page_break(wb, sheet = 1, i = 2, type = "column")
\dontrun{
wb_save(wb, "wb_page_breakExample.xlsx", TRUE)
}
## In Excel: View tab -> Page Break Preview
}
\seealso{
\code{\link[=wb_add_worksheet]{wb_add_worksheet()}}
}
