% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_remove_col_widths}
\alias{wb_remove_col_widths}
\title{Remove column widths from a worksheet}
\usage{
wb_remove_col_widths(wb, sheet, cols)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Indices of columns to remove custom width (if any) from.}
}
\value{
A \code{wbWorkbook} object, invisibly
}
\description{
Remove column widths from a worksheet
}
\examples{
## Create a new workbook
wb <- loadWorkbook(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))

## remove column widths in columns 1 to 20
wb$removeColWidths(1, cols = 1:20)
\dontrun{
wb_save(wb, "removeColWidthsExample.xlsx", overwrite = TRUE)
}
}
\seealso{
\code{\link[=setColWidths]{setColWidths()}}
}
