% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_remove_worksheet}
\alias{wb_remove_worksheet}
\title{Remove a worksheet from a workbook}
\usage{
wb_remove_worksheet(wb, sheet)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}
}
\description{
Remove a worksheet from a Workbook object

Remove a worksheet from a workbook
}
\examples{
## load a workbook
wb <- loadWorkbook(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))

## Remove sheet 2
wb <- wb_remove_worksheet(wb, 2)

## save the modified workbook
\dontrun{
wb_save(wb, "removeWorksheetExample.xlsx", overwrite = TRUE)
}
}
