% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_bookview}
\alias{wb_set_bookview}
\title{Set the workbook position, size and filter}
\usage{
wb_set_bookview(
  wb,
  activeTab = NULL,
  autoFilterDateGrouping = NULL,
  firstSheet = NULL,
  minimized = NULL,
  showHorizontalScroll = NULL,
  showSheetTabs = NULL,
  showVerticalScroll = NULL,
  tabRatio = NULL,
  visibility = NULL,
  windowHeight = NULL,
  windowWidth = NULL,
  xWindow = NULL,
  yWindow = NULL
)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}

\item{activeTab}{activeTab}

\item{autoFilterDateGrouping}{autoFilterDateGrouping}

\item{firstSheet}{firstSheet}

\item{minimized}{minimized}

\item{showHorizontalScroll}{showHorizontalScroll}

\item{showSheetTabs}{showSheetTabs}

\item{showVerticalScroll}{showVerticalScroll}

\item{tabRatio}{tabRatio}

\item{visibility}{visibility}

\item{windowHeight}{windowHeight}

\item{windowWidth}{windowWidth}

\item{xWindow}{xWindow}

\item{yWindow}{yWindow}
}
\value{
The \code{wbWorkbook} object
}
\description{
Get the base font used in the workbook.
}
