% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_cell_style}
\alias{wb_set_cell_style}
\title{get cell style}
\usage{
wb_set_cell_style(wb, sheet, dims, style)
}
\arguments{
\item{sheet}{sheet}

\item{dims}{dims}

\item{style}{style}
}
\description{
get cell style
}
