% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_col_widths}
\alias{wb_set_col_widths}
\title{Set Worksheet column widths}
\usage{
wb_set_col_widths(
  wb,
  sheet = current_sheet(),
  cols,
  widths = 8.43,
  hidden = FALSE
)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object}

\item{sheet}{A name or index of a worksheet, a vector in the case of \code{remove}}

\item{cols}{Indices of cols to set/remove width}

\item{widths}{width to set cols to specified in Excel column width units or "auto" for automatic sizing.
The widths argument is recycled to the length of cols. The default width is 8.43.
Though there is no specific default width for Excel, it depends on the Excel version,
operating system and DPI settings used. Setting it to specific value also is
no guarantee that the output will be of the selected width.}

\item{hidden}{A logical vector to determine which cols are hidden; values
are recycled to the length of \code{cols}.}
}
\description{
This function was renamed to \code{\link[=wb_add_col_widths]{wb_add_col_widths()}}, see documentation there.
}
\keyword{internal}
