% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_properties}
\alias{wb_set_properties}
\title{Modify workbook properties}
\usage{
wb_set_properties(wb, title = NULL, subject = NULL, category = NULL)
}
\arguments{
\item{wb}{A Workbook object}

\item{title, subject, category}{Workbook property, a string.}
}
\value{
A wbWorkbook object, invisibly.
}
\description{
This function is useful for workbooks that are loeaded. It can be used to set the
workbook \code{title}, \code{subject} and \code{category} field. Use \code{\link[=wb_workbook]{wb_workbook()}}
to easily set these properties with a new workbook.
}
\examples{
file <- system.file("extdata", "openxlsx2_example.xlsx", package = "openxlsx2")
wb <- wb_load(file)
wb$title

# Add a title to properties
wb$set_properties(title = "my title")

wb$title

}
\seealso{
\code{\link[=wb_workbook]{wb_workbook()}}
}
