% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_row_heights}
\alias{wb_set_row_heights}
\title{Set worksheet row heights}
\usage{
wb_set_row_heights(
  wb,
  sheet = current_sheet(),
  rows,
  heights = NULL,
  hidden = FALSE
)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}

\item{sheet}{A name or index of a worksheet}

\item{rows}{Indices of rows to set height}

\item{heights}{Heights to set rows to specified in Excel column height units.}

\item{hidden}{Option to hide rows.}
}
\description{
This function was deprecated in favour of \code{\link[=wb_add_row_heights]{wb_add_row_heights()}}
}
\keyword{internal}
