% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_sheet_names}
\alias{wb_set_sheet_names}
\title{Set worksheet names for a workbook}
\usage{
wb_set_sheet_names(wb, old = NULL, new)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}

\item{old}{The name (or index) of the old sheet name. If \code{NULL} will assume
all worksheets are to be renamed.}

\item{new}{The name of the new sheet}
}
\value{
The \link{wbWorkbook} object
}
\description{
Sets the worksheet names for a \link{wbWorkbook} object.
}
\details{
This only changes the sheet name as shown in spreadsheet software
and will not alter it anywhere else. Not in formulas, chart references,
named regions, pivot tables or anywhere else.
}
