% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_sheet_visibility}
\alias{wb_sheet_visibility}
\alias{wb_sheet_visibility<-}
\title{Get/set worksheet visible state}
\usage{
wb_sheet_visibility(wb)

wb_sheet_visibility(wb) <- value
}
\arguments{
\item{wb}{A workbook object}

\item{value}{a logical/character vector the same length as wb_sheet_visibility(wb)}
}
\value{
Character vector of worksheet names.

Vector of "hidden", "visible", "veryHidden"
}
\description{
Get and set worksheet visible state
}
\examples{

wb <- wb_workbook()
wb$add_worksheet(sheet = "S1", visible = FALSE)
wb$add_worksheet(sheet = "S2", visible = TRUE)
wb$add_worksheet(sheet = "S3", visible = FALSE)

wb_sheet_visibility(wb)
wb_sheet_visibility(wb)[1] <- TRUE ## show sheet 1
wb_sheet_visibility(wb)[2] <- FALSE ## hide sheet 2
wb_sheet_visibility(wb)[3] <- "hidden" ## hide sheet 3
wb_sheet_visibility(wb)[3] <- "veryHidden" ## hide sheet 3 from UI
}
