% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_update_table}
\alias{wb_update_table}
\title{update a data_table}
\usage{
wb_update_table(wb, sheet = current_sheet(), dims = "A1", tabname)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{a worksheet}

\item{dims}{cell used as start}

\item{tabname}{a tablename}
}
\description{
update a data_table
}
