% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_ws}
\alias{wb_ws}
\title{little worksheet helper}
\usage{
wb_ws(wb, sheet)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{a worksheet either id or character}
}
\description{
little worksheet helper
}
