% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-color.R
\name{wbColor}
\alias{wbColor}
\alias{wb_color}
\title{R6 class for a Workbook Color}
\usage{
wb_color(
  name = NULL,
  auto = NULL,
  indexed = NULL,
  rgb = NULL,
  theme = NULL,
  tint = NULL
)
}
\arguments{
\item{name}{A name of a color known to R}

\item{auto}{A boolean.}

\item{indexed}{An indexed color values.}

\item{rgb}{A rgb color as ARGB hex value "FF000000".}

\item{theme}{A zero based index referencing a value in the theme.}

\item{tint}{A tint value applied. Range from -1 (dark) to 1 (light).}
}
\value{
a \code{wbColor} object
}
\description{
R6 class for a Workbook Color

R6 class for a Workbook Color

Create a new hyperlink object
}
\details{
A color
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{auto}}{auto}

\item{\code{indexed}}{indexed}

\item{\code{rgb}}{rgb}

\item{\code{theme}}{theme}

\item{\code{tint}}{tint}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wbColor-new}{\code{wbColor$new()}}
\item \href{#method-wbColor-get}{\code{wbColor$get()}}
\item \href{#method-wbColor-to_xml}{\code{wbColor$to_xml()}}
\item \href{#method-wbColor-clone}{\code{wbColor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbColor-new"></a>}}
\if{latex}{\out{\hypertarget{method-wbColor-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbColor} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbColor$new(auto = NULL, indexed = NULL, rgb = NULL, theme = NULL, tint = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{auto}}{A boolean.}

\item{\code{indexed}}{An indexed color values.}

\item{\code{rgb}}{A rgb color as ARGB hex value "FF000000".}

\item{\code{theme}}{A zero based index referencing a value in the theme.}

\item{\code{tint}}{A tint value applied. Range from -1 (dark) to 1 (light).}

\item{\code{name}}{A name of a color known to R}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{wbColor} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbColor-get"></a>}}
\if{latex}{\out{\hypertarget{method-wbColor-get}{}}}
\subsection{Method \code{get()}}{
get as vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbColor$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbColor-to_xml"></a>}}
\if{latex}{\out{\hypertarget{method-wbColor-to_xml}{}}}
\subsection{Method \code{to_xml()}}{
Convert to xml
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbColor$to_xml()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of xml
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbColor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wbColor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbColor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
