% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-comment.R
\name{wbComment}
\alias{wbComment}
\title{R6 class for a Workbook Comments}
\value{
The \code{wbComment} object, invisibly; called for its side effects
}
\description{
R6 class for a Workbook Comments

R6 class for a Workbook Comments
}
\details{
A comment
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{text}}{Comment text}

\item{\code{author}}{The comment author}

\item{\code{style}}{A style for the comment}

\item{\code{visible}}{\code{logical}, if \code{FALSE} is not visible}

\item{\code{width}}{Width of the comment in ... units}

\item{\code{height}}{Height of comment in ... units}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wbComment-new}{\code{wbComment$new()}}
\item \href{#method-wbComment-print}{\code{wbComment$print()}}
\item \href{#method-wbComment-clone}{\code{wbComment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbComment-new"></a>}}
\if{latex}{\out{\hypertarget{method-wbComment-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbComment} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbComment$new(text, author, style, visible = TRUE, width = 2, height = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Comment text}

\item{\code{author}}{The comment author}

\item{\code{style}}{A style for the comment}

\item{\code{visible}}{\code{logical}, if \code{FALSE} is not visible}

\item{\code{width}}{Width of the comment in ... units}

\item{\code{height}}{Height of comment in ... units}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{wbComment} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbComment-print"></a>}}
\if{latex}{\out{\hypertarget{method-wbComment-print}{}}}
\subsection{Method \code{print()}}{
Prints the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbComment$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbComment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wbComment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbComment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
