% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sheet-data.R
\name{wbSheetData}
\alias{wbSheetData}
\alias{new_sheet_data}
\title{R6 class for a Workbook Hyperlink}
\usage{
new_sheet_data()
}
\description{
R6 class for a Workbook Hyperlink

R6 class for a Workbook Hyperlink
}
\details{
A hyperlink
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rows}}{rows}

\item{\code{cols}}{cols}

\item{\code{s}}{s}

\item{\code{r}}{r}

\item{\code{t}}{t}

\item{\code{v}}{v}

\item{\code{f}}{f}

\item{\code{row_attr}}{row_attr}

\item{\code{cc}}{cc}

\item{\code{cc_out}}{cc_out}

\item{\code{style_id}}{style_id}

\item{\code{data_count}}{data_count}

\item{\code{n_elements}}{n_elements}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{wbSheetData$new()}}
\item \href{#method-delete}{\code{wbSheetData$delete()}}
\item \href{#method-write}{\code{wbSheetData$write()}}
\item \href{#method-clone}{\code{wbSheetData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbSheetData} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbSheetData$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{wbSheetData} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbSheetData$delete(rows_in, cols_in, grid_expand)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows_in, cols_in}}{The row and columns to delete}

\item{\code{grid_expand}}{If \code{TRUE} will expand the selection of \code{rows_in}, and
\code{cols_in} to the full set, not just the exact positions (similar to
using \verb{[expand.grid()]})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbSheetData} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbSheetData$write(rows_in, cols_in, t_in, v_in, f_in, any_functions = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows_in, cols_in}}{The row and columns to delete}

\item{\code{t_in, v_in, f_in}}{specifications ???}

\item{\code{any_functions}}{Logical ???}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbSheetData} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbSheetData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
