% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-style.R
\name{wbStyle}
\alias{wbStyle}
\title{R6 class for a Workbook Style}
\description{
R6 class for a Workbook Style

R6 class for a Workbook Style
}
\details{
A style
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fontId}}{fontID}

\item{\code{fontName}}{fontName}

\item{\code{fontColour}}{fontColour}

\item{\code{fontSize}}{fontSize}

\item{\code{fontFamily}}{fontFamily}

\item{\code{fontScheme}}{fontScheme}

\item{\code{fontDecoration}}{fontDecoration}

\item{\code{borderTop}}{borderTop}

\item{\code{borderLeft}}{borderLeft}

\item{\code{borderRight}}{borderRight}

\item{\code{borderBottom}}{borderBottom}

\item{\code{borderTopColour}}{borderTopColour}

\item{\code{borderLeftColour}}{borderLeftColour}

\item{\code{borderRightColour}}{borderRightColour}

\item{\code{borderBottomColour}}{borderBottomColour}

\item{\code{borderDiagonal}}{borderDiagonal}

\item{\code{borderDiagonalColour}}{borderDiagonalColour}

\item{\code{borderDiagonalUp}}{borderDiagonalUp}

\item{\code{borderDiagonalDown}}{borderDiagonalDown}

\item{\code{halign}}{Horizontal alignment}

\item{\code{valign}}{Vertical alignment}

\item{\code{indent}}{Indentation}

\item{\code{textRotation}}{The degree of text rotation}

\item{\code{numFmt}}{Number format id}

\item{\code{fill}}{fill}

\item{\code{wrapText}}{wrapText}

\item{\code{locked}}{locked}

\item{\code{hidden}}{hidden}

\item{\code{xfId}}{xfId}

\item{\code{styleShow}}{styleShow}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{wbStyle$new()}}
\item \href{#method-as.list}{\code{wbStyle$as.list()}}
\item \href{#method-print}{\code{wbStyle$print()}}
\item \href{#method-clone}{\code{wbStyle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbStyle} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbStyle$new(
  fontName = NULL,
  fontSize = NULL,
  fontColour = "none",
  numFmt = "GENERAL",
  border = c("none", "top", "bottom", "left", "right", "all"),
  borderColour = getOption("openxlsx.borderColour", "black"),
  borderStyle = getOption("openxlsx.borderStyle", "thin"),
  bgFill = "none",
  fgFill = "none",
  halign = c("left", "right", "center"),
  valign = c("top", "center", "bottom"),
  textDecoration = c("none", "bold", "strikeout", "italic", "underline", "underline2"),
  wrapText = FALSE,
  textRotation = 0,
  indent = NULL,
  locked = FALSE,
  hidden = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fontName}}{fontName}

\item{\code{fontSize}}{fontSize}

\item{\code{fontColour}}{fontColour}

\item{\code{numFmt}}{Number format id}

\item{\code{border}}{Border}

\item{\code{borderColour}}{border colour}

\item{\code{borderStyle}}{border style}

\item{\code{bgFill}}{Background fill}

\item{\code{fgFill}}{Foreground fill}

\item{\code{halign}}{Horizontal alignment}

\item{\code{valign}}{Vertical alignment}

\item{\code{textDecoration}}{text decorations}

\item{\code{wrapText}}{If \code{TRUE} will wrap text}

\item{\code{textRotation}}{The degree of text rotation}

\item{\code{indent}}{indent}

\item{\code{locked}}{locked}

\item{\code{hidden}}{hidden}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{wbStyle} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.list"></a>}}
\if{latex}{\out{\hypertarget{method-as.list}{}}}
\subsection{Method \code{as.list()}}{
Convert the \code{wbStyle} object to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbStyle$as.list(include_null = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_null}}{if \code{TRUE} will include values that contain \code{NULL},
otherwise they are ommited from the list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list} of fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the \verb{wbStyle object}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbStyle$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not implemented}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{wbStyle} object, invisibly; called for its side effects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbStyle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
