% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{write_file}
\alias{write_file}
\title{write xml file}
\usage{
write_file(head = "", body = "", tail = "", fl = "")
}
\arguments{
\item{head}{head part of xml}

\item{body}{body part of xml}

\item{tail}{tail part of xml}

\item{fl}{file name with full path}
}
\description{
brings the added benefit of xml checking
}
