% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{ws_grid_lines}
\alias{ws_grid_lines}
\title{Set worksheet gridlines to show or hide.}
\usage{
ws_grid_lines(wb, sheet, show_grid_lines = FALSE)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{show_grid_lines}{A logical. If \code{FALSE}, grid lines are hidden.}
}
\description{
Set worksheet gridlines to show or hide.
}
\examples{
wb <- wb_load(file = system.file("extdata", "loadExample.xlsx", package = "openxlsx2"))
names(wb) ## list worksheets in workbook
ws_grid_lines(wb, 1, show_grid_lines = FALSE)
ws_grid_lines(wb, "testing", show_grid_lines = FALSE)
\dontrun{
wb_save(wb, "ws_grid_linesExample.xlsx", overwrite = TRUE)
}
}
