% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_add_child}
\alias{xml_add_child}
\title{append xml child to node}
\usage{
xml_add_child(xml_node, xml_child, pointer = FALSE, escapes = FALSE)
}
\arguments{
\item{xml_node}{xml_node}

\item{xml_child}{xml_child}

\item{pointer}{pointer}

\item{escapes}{escapes}
}
\description{
append xml child to node
}
\examples{
xml_node <- "<node><child1/><child2/></node>"
xml_child <- "<new_child/>"

xml_add_child(xml_node, xml_child)
}
