% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_attribute}
\alias{xml_attribute}
\title{xml_attribute}
\usage{
xml_attribute(
  xml,
  level1 = NULL,
  level2 = NULL,
  level3 = NULL,
  level4 = NULL,
  level5 = NULL,
  level6 = NULL
)
}
\arguments{
\item{xml}{something xml}

\item{level1}{to please check}

\item{level2}{to please check}

\item{level3}{to please check}

\item{level4}{to please check}

\item{level5}{to please check}

\item{level6}{to please check}
}
\description{
xml_attribute
}
\examples{

  x <- read_xml("<a a=\"1\" b=\"2\">1</a>")
  xml_attribute(x, "a")

  x <- read_xml("<a><b r=\"1\">2</b></a>")
  xml_attribute(x, "a", "b")
  x <- read_xml("<a a=\"1\" b=\"2\">1</a>")
  xml_attribute(x, "a")

  x <- read_xml("<b><a a=\"1\" b=\"2\"/></b>")
  xml_attribute(x, "b", "a")
}
