% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_child_mod}
\alias{xml_child_mod}
\title{adds or updates attribute(s) in children of existing xml node}
\usage{
xml_child_mod(
  xml_node,
  xml_child,
  xml_attributes = NULL,
  xml_children = NULL,
  ...
)
}
\arguments{
\item{xml_node}{some valid xml_node}

\item{xml_child}{a name of a child node that should be modified}

\item{xml_attributes}{R vector of named attributes}

\item{xml_children}{R character vector children attached to the xml_node}

\item{...}{additional attributes passed to function}
}
\description{
adds or updates attribute(s) in children of existing xml node
}
\details{
If a named attribute in \code{xml_attributes} is "" remove the attribute
from the node.
If \code{xml_attributes} contains a named entry found in the xml node, it is
updated else it is added as attribute.
}
\examples{
xml <- "<foo><bar a=\"0\"/></foo>"
xml_child_mod(xml, xml_child = "bar", xml_children = "<openxlsx2/>")
xml_child_mod(xml, xml_child = "baz", xml_children = "<openxlsx2/>")

xml_child_mod(xml, xml_child = "bar", xml_attributes = c(a = "1", b = "2"))
xml_child_mod(xml, xml_child = "baz", xml_attributes = c(a = "1", b = "2"))

}
\seealso{
\code{\link[=xml_attr_mod]{xml_attr_mod()}}
}
