% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_order_children}
\alias{xml_order_children}
\title{order xml children in node}
\usage{
xml_order_children(xml_node, level, order, pointer = FALSE, ...)
}
\arguments{
\item{xml_node}{an xml structure}

\item{level}{the xml root}

\item{order}{the wanted order as numeric}

\item{pointer}{pointer}

\item{...}{additional arguments passed to \code{read_xml()}}
}
\description{
order xml children in node
}
