% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_rm_child}
\alias{xml_rm_child}
\title{remove xml child to node}
\usage{
xml_rm_child(xml_node, xml_child, level, which = 0, pointer = FALSE, ...)
}
\arguments{
\item{xml_node}{xml_node}

\item{xml_child}{xml_child}

\item{level}{optional level, if missing the first child is picked}

\item{which}{optional index which node to remove, if multiple are available. Default disabled all will be removed}

\item{pointer}{pointer}

\item{...}{additional arguments passed to \code{read_xml()}}
}
\description{
remove xml child to node
}
\examples{
xml_node <- "<a><b><c><d/></c></b><c/></a>"
xml_child <- "c"

xml_rm_child(xml_node, xml_child)

xml_rm_child(xml_node, xml_child, level = c("b"))

xml_rm_child(xml_node, "d", level = c("b", "c"))

}
