% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illegal-characters.R
\name{clean_worksheet_name}
\alias{clean_worksheet_name}
\title{Clean worksheet name}
\usage{
clean_worksheet_name(x, replacement = " ")
}
\arguments{
\item{x}{A vector, coerced to \code{character}}

\item{replacement}{A single value to replace illegal characters by.}
}
\value{
x with bad characters removed
}
\description{
Cleans a worksheet name by removing legal characters.
}
\details{
Illegal characters are considered \verb{\\}, \code{/}, \verb{?}, \code{*}, \code{:}, \code{[}, and
\verb{]}.  These must be intentionally removed from worksheet names prior to
creating a new worksheet.
}
