% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{cleanSheet}
\alias{cleanSheet}
\title{clean sheet (remove all values)}
\usage{
cleanSheet(
  wb,
  sheet,
  numbers = TRUE,
  characters = TRUE,
  styles = TRUE,
  merged_cells = TRUE
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet to clean}

\item{numbers}{remove all numbers}

\item{characters}{remove all characters}

\item{styles}{remove all styles}

\item{merged_cells}{remove all merged_cells}
}
\description{
clean sheet (remove all values)
}
