% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R, R/wb_functions.R
\name{cleanup}
\alias{cleanup}
\alias{wb_clean_sheet}
\alias{delete_data}
\title{clean sheet (remove all values)}
\usage{
wb_clean_sheet(
  wb,
  sheet = current_sheet(),
  dims = NULL,
  numbers = TRUE,
  characters = TRUE,
  styles = TRUE,
  merged_cells = TRUE
)

delete_data(wb, sheet, cols, rows)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet to clean}

\item{dims}{spreadsheet dimensions (optional)}

\item{numbers}{remove all numbers}

\item{characters}{remove all characters}

\item{styles}{remove all styles}

\item{merged_cells}{remove all merged_cells}

\item{cols}{numeric column vector}

\item{rows}{numeric row vector}
}
\description{
clean sheet (remove all values)
}
