% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{col2int}
\alias{col2int}
\title{Convert spreadsheet column to integer}
\usage{
col2int(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
An integer column label (or \code{NULL} if \code{x} is \code{NULL})
}
\description{
Converts a spreadsheet column in \code{A1} notation to an integer.
\code{"A"} is \code{1}, \code{"B"} is \code{2}, ..., \code{"Z"} is \code{26} and \code{"AA"} is \code{27}.
}
\examples{
col2int(LETTERS)
}
