% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{col2int}
\alias{col2int}
\title{Convert Excel column to integer}
\usage{
col2int(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
An integer column label (or \code{NULL} if \code{x} is \code{NULL})
}
\description{
Converts an Excel column label to an integer.
}
\examples{
col2int(LETTERS)
}
