% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{convertToExcelDate}
\alias{convertToExcelDate}
\title{convert back to ExcelDate}
\usage{
convertToExcelDate(df, date1904 = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{date1904}{take different origin}
}
\description{
convert back to ExcelDate
}
