% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{convertToExcelDate}
\alias{convertToExcelDate}
\title{Convert to Excel data}
\arguments{
\item{df}{dataframe}

\item{date1904}{take different origin}
}
\description{
Use \code{\link[=convert_to_excel_date]{convert_to_excel_date()}}.
}
\keyword{internal}
