% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{convertToExcelDate}
\alias{convertToExcelDate}
\title{convert back to ExcelDate}
\usage{
convertToExcelDate(df, date1904 = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{date1904}{take different origin}
}
\description{
convert back to ExcelDate
}
\examples{
 xlsxFile <- system.file("extdata", "readTest.xlsx", package = "openxlsx2")
 wb1 <- wb_load(xlsxFile)
 df <- wb_to_df(wb1)
 # conversion is done on dataframes only
 convertToExcelDate(df = df["Var5"], date1904 = FALSE)
}
