% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_border}
\alias{create_border}
\title{create fill}
\usage{
create_border(
  diagonalDown = "",
  diagonalUp = "",
  outline = "",
  bottom = "",
  bottom_color = "",
  diagonal = "",
  diagonal_color = "",
  end = "",
  horizontal = "",
  left = "",
  left_color = "",
  right = "",
  right_color = "",
  start = "",
  top = "",
  top_color = "",
  vertical = ""
)
}
\arguments{
\item{diagonalDown}{x}

\item{diagonalUp}{x}

\item{outline}{x}

\item{bottom}{X}

\item{bottom_color}{X}

\item{diagonal}{X}

\item{diagonal_color}{X,}

\item{end}{x,}

\item{horizontal}{x}

\item{left}{x}

\item{left_color}{x}

\item{right}{x}

\item{right_color}{x}

\item{start}{x}

\item{top}{x}

\item{top_color}{x}

\item{vertical}{x}
}
\description{
create fill
}
