% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_border}
\alias{create_border}
\title{create border}
\usage{
create_border(
  diagonalDown = "",
  diagonalUp = "",
  outline = "",
  bottom = NULL,
  bottom_color = NULL,
  diagonal = NULL,
  diagonal_color = NULL,
  end = "",
  horizontal = "",
  left = NULL,
  left_color = NULL,
  right = NULL,
  right_color = NULL,
  start = "",
  top = NULL,
  top_color = NULL,
  vertical = ""
)
}
\arguments{
\item{diagonalDown}{x}

\item{diagonalUp}{x}

\item{outline}{x}

\item{bottom}{X}

\item{bottom_color}{X}

\item{diagonal}{X}

\item{diagonal_color}{X,}

\item{end}{x,}

\item{horizontal}{x}

\item{left}{x}

\item{left_color}{x}

\item{right}{x}

\item{right_color}{x}

\item{start}{x}

\item{top}{x}

\item{top_color}{x}

\item{vertical}{x}
}
\description{
Border styles can any of the following: "thin", "thick", "slantDashDot", "none", "mediumDashed", "mediumDashDot", "medium", "hair", "double", "dotted", "dashed", "dashedDotDot", "dashDot"
Border colors are of the following type: c(rgb="FF000000")
}
