% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_cell_style}
\alias{create_cell_style}
\title{create_cell_style}
\usage{
create_cell_style(
  borderId = "",
  fillId = "",
  fontId = "",
  numFmtId = "",
  pivotButton = "",
  quotePrefix = "",
  xfId = "",
  horizontal = "",
  indent = "",
  justifyLastLine = "",
  readingOrder = "",
  relativeIndent = "",
  shrinkToFit = "",
  textRotation = "",
  vertical = "",
  wrapText = "",
  extLst = "",
  hidden = "",
  locked = "",
  ...
)
}
\arguments{
\item{borderId}{dummy}

\item{fillId}{dummy}

\item{fontId}{dummy}

\item{numFmtId}{a numFmt ID for a builtin style}

\item{pivotButton}{dummy}

\item{quotePrefix}{dummy}

\item{xfId}{dummy}

\item{horizontal}{alignment can be "", "center", "right"}

\item{indent}{dummy}

\item{justifyLastLine}{dummy}

\item{readingOrder}{dummy}

\item{relativeIndent}{dummy}

\item{shrinkToFit}{dummy}

\item{textRotation}{dummy}

\item{vertical}{alignment can be "", "center", "right"}

\item{wrapText}{dummy}

\item{extLst}{dummy}

\item{hidden}{dummy}

\item{locked}{dummy}

\item{...}{reserved for additional arguments}
}
\description{
create_cell_style
}
\details{
\tabular{ll}{
   "ID" \tab "numFmt" \cr
   "0" \tab "General" \cr
   "1" \tab "0" \cr
   "2" \tab "0.00" \cr
   "3" \tab "#,##0" \cr
   "4" \tab "#,##0.00" \cr
   "9" \tab "0\%" \cr
   "10" \tab "0.00\%" \cr
   "11" \tab "0.00E+00" \cr
   "12" \tab "# ?/?" \cr
   "13" \tab "# ??/??" \cr
   "14" \tab "mm-dd-yy" \cr
   "15" \tab "d-mmm-yy" \cr
   "16" \tab "d-mmm" \cr
   "17" \tab "mmm-yy" \cr
   "18" \tab "h:mm AM/PM" \cr
   "19" \tab "h:mm:ss AM/PM" \cr
   "20" \tab "h:mm" \cr
   "21" \tab "h:mm:ss" \cr
   "22" \tab "m/d/yy h:mm" \cr
   "37" \tab "#,##0 ;(#,##0)" \cr
   "38" \tab "#,##0 ;[Red](#,##0)" \cr
   "39" \tab "#,##0.00;(#,##0.00)" \cr
   "40" \tab "#,##0.00;[Red](#,##0.00)" \cr
   "45" \tab "mm:ss" \cr
   "46" \tab "[h]:mm:ss" \cr
   "47" \tab "mmss.0" \cr
   "48" \tab "##0.0E+0" \cr
   "49" \tab "@" \cr
}
}
