% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_colors_xml}
\alias{create_colors_xml}
\alias{create_colours_xml}
\title{create custom color xml schemes
To be used with \code{\link[=wb_set_base_colors]{wb_set_base_colors()}}}
\usage{
create_colors_xml(name = "Base R", dark = NULL, accent = NULL, link = NULL)
}
\arguments{
\item{name}{the color name}

\item{dark}{four colors: dark, light, brighter dark, darker light}

\item{accent}{six accent colors}

\item{link}{two link colors: link and visited link}
}
\description{
create custom color xml schemes
To be used with \code{\link[=wb_set_base_colors]{wb_set_base_colors()}}
}
\examples{
colors <- create_colors_xml()
wb <- wb_workbook()$add_worksheet()$set_base_colors(xml = colors)
}
