% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{create_comment}
\alias{create_comment}
\title{Create a comment}
\usage{
create_comment(
  text,
  author = Sys.info()[["user"]],
  style = NULL,
  visible = TRUE,
  width = 2,
  height = 4
)
}
\arguments{
\item{text}{Comment text. Character vector or a \code{\link[=fmt_txt]{fmt_txt()}} string.}

\item{author}{A string, by default, will use "user"}

\item{style}{A Style object or list of style objects the same length as comment vector.}

\item{visible}{Default: \code{TRUE}. Is the comment visible by default?}

\item{width}{Textbox integer width in number of cells}

\item{height}{Textbox integer height in number of cells}
}
\value{
a \code{wbComment} object
}
\description{
Use \code{\link[=wb_comment]{wb_comment()}} in new code. See \link{openxlsx2-deprecated}
}
\keyword{internal}
