% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_fill}
\alias{create_fill}
\title{Create fill pattern}
\usage{
create_fill(
  gradientFill = "",
  patternType = "",
  bgColor = NULL,
  fgColor = NULL,
  ...
)
}
\arguments{
\item{gradientFill}{complex fills}

\item{patternType}{various: default is "none", but also "solid", or a color like "gray125"}

\item{bgColor}{hex8 color with alpha, red, green, blue only for patternFill}

\item{fgColor}{hex8 color with alpha, red, green, blue only for patternFill}

\item{...}{...}
}
\description{
Create fill pattern
}
\seealso{
\code{\link[=wb_add_fill]{wb_add_fill()}}

Other style creating functions: 
\code{\link{create_border}()},
\code{\link{create_cell_style}()},
\code{\link{create_colors_xml}()},
\code{\link{create_dxfs_style}()},
\code{\link{create_font}()},
\code{\link{create_numfmt}()},
\code{\link{create_tablestyle}()}
}
\concept{style creating functions}
