% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_fill}
\alias{create_fill}
\title{create fill}
\usage{
create_fill(
  gradientFill = "",
  patternType = "",
  bgColor = NULL,
  fgColor = NULL,
  ...
)
}
\arguments{
\item{gradientFill}{complex fills}

\item{patternType}{various default is "none", but also "solid", or a color like "gray125"}

\item{bgColor}{hex8 color with alpha, red, green, blue only for patternFill}

\item{fgColor}{hex8 color with alpha, red, green, blue only for patternFill}

\item{...}{...}
}
\description{
create fill
}
