% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_tablestyle}
\alias{create_tablestyle}
\title{create tableStyle}
\usage{
create_tablestyle(
  name,
  whole_table = NULL,
  header_row = NULL,
  total_row = NULL,
  first_column = NULL,
  last_column = NULL,
  first_row_stripe = NULL,
  second_row_stripe = NULL,
  first_column_stripe = NULL,
  second_column_stripe = NULL,
  first_header_cell = NULL,
  last_header_cell = NULL,
  first_total_cell = NULL,
  last_total_cell = NULL,
  ...
)
}
\arguments{
\item{name}{name}

\item{whole_table}{wholeTable}

\item{header_row}{headerRow}

\item{total_row}{totalRow}

\item{first_column}{firstColumn}

\item{last_column}{lastColumn}

\item{first_row_stripe}{firstRowStripe}

\item{second_row_stripe}{secondRowStripe}

\item{first_column_stripe}{firstColumnStripe}

\item{second_column_stripe}{secondColumnStripe}

\item{first_header_cell}{firstHeaderCell}

\item{last_header_cell}{lastHeaderCell}

\item{first_total_cell}{firstTotalCell}

\item{last_total_cell}{lastTotalCell}

\item{...}{additional arguments}
}
\description{
Create a custom table style.
This function is for expert use only. Use other styling functions instead.
}
