% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_styles.R
\name{create_tablestyle}
\alias{create_tablestyle}
\title{create tableStyle}
\usage{
create_tablestyle(
  name,
  wholeTable = NULL,
  headerRow = NULL,
  totalRow = NULL,
  firstColumn = NULL,
  lastColumn = NULL,
  firstRowStripe = NULL,
  secondRowStripe = NULL,
  firstColumnStripe = NULL,
  secondColumnStripe = NULL,
  firstHeaderCell = NULL,
  lastHeaderCell = NULL,
  firstTotalCell = NULL,
  lastTotalCell = NULL
)
}
\arguments{
\item{name}{name}

\item{wholeTable}{wholeTable}

\item{headerRow}{headerRow}

\item{totalRow}{totalRow}

\item{firstColumn}{firstColumn}

\item{lastColumn}{lastColumn}

\item{firstRowStripe}{firstRowStripe}

\item{secondRowStripe}{secondRowStripe}

\item{firstColumnStripe}{firstColumnStripe}

\item{secondColumnStripe}{secondColumnStripe}

\item{firstHeaderCell}{firstHeaderCell}

\item{lastHeaderCell}{lastHeaderCell}

\item{firstTotalCell}{firstTotalCell}

\item{lastTotalCell}{lastTotalCell}
}
\description{
create tableStyle
}
