% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dims_helper}
\alias{dims_helper}
\alias{dims_to_rowcol}
\alias{rowcol_to_dims}
\title{dims helpers}
\usage{
dims_to_rowcol(x, as_integer = FALSE)

rowcol_to_dims(row, col)
}
\arguments{
\item{x}{a dimension object "A1" or "A1:A1"}

\item{as_integer}{optional if the output should be returned as integer}

\item{row}{a numeric vector of rows}

\item{col}{a numeric or character vector of cols}
}
\description{
Internal helpers to (de)construct a dims argument from/to a row
and column vector. Exported for user convenience.
}
\examples{
dims_to_rowcol("A1:J10")
rowcol_to_dims(1:10, 1:10)
}
