% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{freezePane}
\alias{freezePane}
\alias{wb_freeze_pane}
\title{Freeze a worksheet pane}
\usage{
wb_freeze_pane(
  wb,
  sheet,
  firstActiveRow = NULL,
  firstActiveCol = NULL,
  firstRow = FALSE,
  firstCol = FALSE
)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{firstActiveRow}{Top row of active region}

\item{firstActiveCol}{Furthest left column of active region}

\item{firstRow}{If \code{TRUE}, freezes the first row (equivalent to firstActiveRow = 2)}

\item{firstCol}{If \code{TRUE}, freezes the first column (equivalent to firstActiveCol = 2)}
}
\description{
Freeze a worksheet pane
}
\examples{
## Create a new workbook
wb <- wb_workbook("Kenshin")

## Add some worksheets
wb$addWorksheet("Sheet 1")
wb$addWorksheet("Sheet 2")
wb$addWorksheet("Sheet 3")
wb$addWorksheet("Sheet 4")

## Freeze Panes
wb$freezePanes("Sheet 1", firstActiveRow = 5, firstActiveCol = 3)
wb$freezePanes("Sheet 2", firstCol = TRUE) ## shortcut to firstActiveCol = 2
wb$freezePanes(3, firstRow = TRUE) ## shortcut to firstActiveRow = 2
wb$freezePanes(4, firstActiveRow = 1, firstActiveCol = "D")

## Save workbook
\dontrun{
wb_save(wb, "freezePaneExample.xlsx", overwrite = TRUE)
}
}
