% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_group_cols}
\alias{wb_group_cols}
\alias{ungroupColumns}
\alias{grouping}
\alias{wb_group_rows}
\alias{ungroupRows}
\title{Group Rows and Columns}
\usage{
wb_group_cols(wb, sheet, cols, collapsed = FALSE, levels = NULL)

ungroupColumns(wb, sheet, cols)

wb_group_rows(wb, sheet, rows, collapsed = FALSE, levels = NULL)

ungroupRows(wb, sheet, rows)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Indices of columns to ungroup}

\item{collapsed}{Logical vector. If TRUE the grouped columns are collapsed. Defaults to FALSE}

\item{levels}{levels?}

\item{rows}{Indices of rows to ungroup}
}
\description{
Group a selection of rows or cols
}
\details{
Group columns together, with the option to hide them.

If column was previously hidden, it will now be shown

If row was previously hidden, it will now be shown
}
\examples{

# create matrix
t1 <- AirPassengers
t2 <- do.call(cbind, split(t1, cycle(t1)))
dimnames(t2) <- dimnames(.preformat.ts(t1))

wb <- wb_workbook()
wb <- wb_add_worksheet(wb, "AirPass")
writeData(wb, "AirPass", t2, rowNames = TRUE)

# groups will always end on/show the last row. in the example 1950, 1955, and 1960
wb <- wb_group_rows(wb, "AirPass", 2:3, collapsed = TRUE) # group years < 1950
wb <- wb_group_rows(wb, "AirPass", 4:8, collapsed = TRUE) # group years 1951-1955
wb <- wb_group_rows(wb, "AirPass", 9:13)                  # group years 1956-1960

wb$createCols("AirPass", 13)

wb <- wb_group_cols(wb, "AirPass", 2:4, collapsed = TRUE)
wb <- wb_group_cols(wb, "AirPass", 5:7, collapsed = TRUE)
wb <- wb_group_cols(wb, "AirPass", 8:10, collapsed = TRUE)
wb <- wb_group_cols(wb, "AirPass", 11:13)

}
