% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{int2col}
\alias{int2col}
\title{Convert integer to spreadsheet column}
\usage{
int2col(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Converts an integer to a spreadsheet column in \code{A1} notation.
\code{1} is \code{"A"}, \code{2} is \code{"B"}, ..., \code{26} is \code{"Z"} and \code{27} is \code{"AA"}.
}
\examples{
int2col(1:10)
}
