% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names.wbWorkbook}
\alias{names.wbWorkbook}
\alias{names<-.wbWorkbook}
\title{get or set worksheet names
get or set worksheet names}
\usage{
\method{names}{wbWorkbook}(x)

\method{names}{wbWorkbook}(x) <- value
}
\arguments{
\item{x}{A \code{wbWorkbook} object}

\item{value}{a character vector the same length as \code{x}}
}
\description{
get or set worksheet names
get or set worksheet names
}
