% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2.R
\name{openxlsx2-deprecated}
\alias{openxlsx2-deprecated}
\title{Deprecated Functions in Package \emph{openxlsx2}}
\description{
These functions are provided for compatibility with older versions of \code{openxlsx2}, and may be defunct as soon as the next release.
}
\details{
\code{convertToExcelDate()} -> \code{convert_to_excel_date()}
}
\seealso{
\link{.Deprecated}
}
