% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2-package.R
\name{openxlsx2-deprecated}
\alias{openxlsx2-deprecated}
\title{Deprecated Functions in Package \emph{openxlsx2}}
\description{
These functions are provided for compatibility with older versions of \code{openxlsx2}, and may be defunct as soon as the next release.
}
\details{
\itemize{
\item \code{\link[=convertToExcelDate]{convertToExcelDate()}} -> \code{\link[=convert_to_excel_date]{convert_to_excel_date()}}
\item \code{\link[=wb_grid_lines]{wb_grid_lines()}} -> \code{\link[=wb_set_grid_lines]{wb_set_grid_lines()}}
}
}
\seealso{
\link{.Deprecated}
}
