% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-style_mgr.R
\name{style_mgr}
\alias{style_mgr}
\title{style manager}
\description{
style manager

style manager
}
\examples{
xlsxFile <- system.file("extdata", "oxlsx2_sheet.xlsx", package = "openxlsx2")
wb <- wb_load(xlsxFile)

# ## start style mgr
# style <- style_mgr$new(wb)
# style$initialize(wb)

# wb$styles_mgr$get_numfmt() |> print()
# wb$styles_mgr$next_numfmt_id() |> print()
# wb$styles_mgr$get_numfmt_id("numFmt-166")

# create new number format
new_numfmt <- create_numfmt(numFmtId = wb$styles_mgr$next_numfmt_id(), formatCode = "#,#")

# add it via stylemgr
wb$styles_mgr$add(new_numfmt, "test")

## get numfmts (invisible)
# z <- wb$styles_mgr$get_numfmt()
# z
wb$styles_mgr$styles$numFmts

## create and add huge font
new_huge_font <- create_font(sz = "20", name = "Arial", b = "1",
                             color = wb_colour(hex = "FFFFFFFF"))
wb$styles_mgr$add(new_huge_font, "arial_huge")

## create another font
new_font <- create_font(name = "Arial")
wb$styles_mgr$add(new_font, "arial")

## create and add new fill
new_fill <- create_fill(patternType = "solid", fgColor = wb_colour(hex = "FF00224B"))
wb$styles_mgr$add(new_fill, "blue")

# create new style with numfmt enabled
head_xf <- create_cell_style(
  horizontal = "center",
  textRotation = "45",
  numFmtId = "0",
  fontId = wb$styles_mgr$get_font_id("arial_huge"),
  fillId = wb$styles_mgr$get_fill_id("blue")
)

new_xf <- create_cell_style(
  numFmtId = wb$styles_mgr$get_numfmt_id("test"),
  fontId = wb$styles_mgr$get_font_id("arial")
)

## add new styles
wb$styles_mgr$add(head_xf, "head_xf")
wb$styles_mgr$add(new_xf, "new_xf")

## get cell style ids (invisible)
# z <- wb$styles_mgr$get_xf()

## get cell style id
# wb$styles_mgr$get_xf_id("new_xf")

 ## assign styles to cells
wb$set_cell_style("SUM", "B3:I3", wb$styles_mgr$get_xf_id("head_xf"))
wb$set_cell_style("SUM", "C7:C16", wb$styles_mgr$get_xf_id("new_xf"))
# wb_open(wb)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numfmt}}{numfmt-ids}

\item{\code{font}}{font-ids}

\item{\code{fill}}{fill-ids}

\item{\code{border}}{border-ids}

\item{\code{xf}}{xf-ids}

\item{\code{dxf}}{dxf-ids}

\item{\code{styles}}{styles as xml}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wbStylesMgr-new}{\code{style_mgr$new()}}
\item \href{#method-wbStylesMgr-get_numfmt}{\code{style_mgr$get_numfmt()}}
\item \href{#method-wbStylesMgr-get_font}{\code{style_mgr$get_font()}}
\item \href{#method-wbStylesMgr-get_fill}{\code{style_mgr$get_fill()}}
\item \href{#method-wbStylesMgr-get_border}{\code{style_mgr$get_border()}}
\item \href{#method-wbStylesMgr-get_xf}{\code{style_mgr$get_xf()}}
\item \href{#method-wbStylesMgr-get_dxf}{\code{style_mgr$get_dxf()}}
\item \href{#method-wbStylesMgr-get_numfmt_id}{\code{style_mgr$get_numfmt_id()}}
\item \href{#method-wbStylesMgr-get_font_id}{\code{style_mgr$get_font_id()}}
\item \href{#method-wbStylesMgr-get_fill_id}{\code{style_mgr$get_fill_id()}}
\item \href{#method-wbStylesMgr-get_border_id}{\code{style_mgr$get_border_id()}}
\item \href{#method-wbStylesMgr-get_xf_id}{\code{style_mgr$get_xf_id()}}
\item \href{#method-wbStylesMgr-get_dxf_id}{\code{style_mgr$get_dxf_id()}}
\item \href{#method-wbStylesMgr-next_numfmt_id}{\code{style_mgr$next_numfmt_id()}}
\item \href{#method-wbStylesMgr-next_font_id}{\code{style_mgr$next_font_id()}}
\item \href{#method-wbStylesMgr-next_fill_id}{\code{style_mgr$next_fill_id()}}
\item \href{#method-wbStylesMgr-next_border_id}{\code{style_mgr$next_border_id()}}
\item \href{#method-wbStylesMgr-next_xf_id}{\code{style_mgr$next_xf_id()}}
\item \href{#method-wbStylesMgr-next_dxf_id}{\code{style_mgr$next_dxf_id()}}
\item \href{#method-wbStylesMgr-add}{\code{style_mgr$add()}}
\item \href{#method-wbStylesMgr-clone}{\code{style_mgr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-new"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbStylesMgr} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$new(
  numfmt = NA,
  font = NA,
  fill = NA,
  border = NA,
  xf = NA,
  dxf = NA,
  styles = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numfmt}}{numfmt}

\item{\code{font}}{font}

\item{\code{fill}}{fill}

\item{\code{border}}{border}

\item{\code{xf}}{xf}

\item{\code{dxf}}{dxf}

\item{\code{styles}}{styles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{wbStylesMgr} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_numfmt"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_numfmt}{}}}
\subsection{Method \code{get_numfmt()}}{
get numfmt ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_numfmt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_font"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_font}{}}}
\subsection{Method \code{get_font()}}{
get font ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_font()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_fill"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_fill}{}}}
\subsection{Method \code{get_fill()}}{
get fill ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_fill()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_border"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_border}{}}}
\subsection{Method \code{get_border()}}{
get border ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_border()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_xf"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_xf}{}}}
\subsection{Method \code{get_xf()}}{
get xf ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_xf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_dxf"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_dxf}{}}}
\subsection{Method \code{get_dxf()}}{
get dxf ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_dxf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_numfmt_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_numfmt_id}{}}}
\subsection{Method \code{get_numfmt_id()}}{
get numfmt id by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_numfmt_id(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_font_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_font_id}{}}}
\subsection{Method \code{get_font_id()}}{
get font id by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_font_id(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_fill_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_fill_id}{}}}
\subsection{Method \code{get_fill_id()}}{
get fill id by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_fill_id(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_border_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_border_id}{}}}
\subsection{Method \code{get_border_id()}}{
get border id by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_border_id(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_xf_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_xf_id}{}}}
\subsection{Method \code{get_xf_id()}}{
get xf id by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_xf_id(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-get_dxf_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-get_dxf_id}{}}}
\subsection{Method \code{get_dxf_id()}}{
get dxf id by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$get_dxf_id(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-next_numfmt_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-next_numfmt_id}{}}}
\subsection{Method \code{next_numfmt_id()}}{
get next numfmt id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$next_numfmt_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-next_font_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-next_font_id}{}}}
\subsection{Method \code{next_font_id()}}{
get next font id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$next_font_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-next_fill_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-next_fill_id}{}}}
\subsection{Method \code{next_fill_id()}}{
get next fill id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$next_fill_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-next_border_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-next_border_id}{}}}
\subsection{Method \code{next_border_id()}}{
get next border id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$next_border_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-next_xf_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-next_xf_id}{}}}
\subsection{Method \code{next_xf_id()}}{
get next xf id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$next_xf_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-next_dxf_id"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-next_dxf_id}{}}}
\subsection{Method \code{next_dxf_id()}}{
get next xf id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$next_dxf_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-add"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-add}{}}}
\subsection{Method \code{add()}}{
add entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$add(style, style_name, skip_duplicates = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{new_style}

\item{\code{style_name}}{a unique name identifying the style}

\item{\code{skip_duplicates}}{should duplicates be added?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbStylesMgr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wbStylesMgr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{style_mgr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
