% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_border}
\alias{wb_add_border}
\title{add border for cell region}
\usage{
wb_add_border(
  wb,
  sheet,
  dims = "A1",
  bottom_color = c(rgb = "FF000000"),
  left_color = c(rgb = "FF000000"),
  right_color = c(rgb = "FF000000"),
  top_color = c(rgb = "FF000000"),
  bottom_border = "thin",
  left_border = "thin",
  right_border = "thin",
  top_border = "thin"
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{a worksheet}

\item{dims}{dimensions on the worksheet e.g. "A1", "A1:A5", "A1:H5"}

\item{bottom_color, left_color, right_color, top_color}{a color, either something openxml knows or some RGB color}

\item{left_border, right_border, top_border, bottom_border}{the border style, if NULL no border is drawn. See create_border for possible border styles}
}
\description{
wb wrapper to create borders for cell region
}
\seealso{
create_border
}
