% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_chart_xml}
\alias{wb_add_chart_xml}
\title{dummy function to add a chart to an existing workbook
currently only a barplot is possible}
\usage{
wb_add_chart_xml(wb, sheet = current_sheet(), xml, dims = NULL)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{the sheet on which the graph will appear}

\item{xml}{chart xml}

\item{dims}{the dimensions where the sheet will appear}
}
\description{
dummy function to add a chart to an existing workbook
currently only a barplot is possible
}
