% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_chart_xml}
\alias{wb_add_chart_xml}
\title{add a chart xml to a workbook}
\usage{
wb_add_chart_xml(
  wb,
  sheet = current_sheet(),
  dims = NULL,
  xml,
  col_offset = 0,
  row_offset = 0,
  ...
)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{the sheet on which the graph will appear}

\item{dims}{the dimensions where the sheet will appear}

\item{xml}{chart xml}

\item{col_offset, row_offset}{positioning}

\item{...}{additional arguments}
}
\description{
add a chart xml to a workbook
}
\seealso{
\code{\link[=wb_add_drawing]{wb_add_drawing()}} \code{\link[=wb_add_image]{wb_add_image()}} \code{\link[=wb_add_mschart]{wb_add_mschart()}} \code{\link[=wb_add_plot]{wb_add_plot()}}
}
