% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_ignore_error}
\alias{wb_add_ignore_error}
\title{Ignore error on worksheet}
\usage{
wb_add_ignore_error(
  wb,
  sheet = current_sheet(),
  dims = "A1",
  calculated_column = FALSE,
  empty_cell_reference = FALSE,
  eval_error = FALSE,
  formula = FALSE,
  formula_range = FALSE,
  list_data_validation = FALSE,
  number_stored_as_text = FALSE,
  two_digit_text_year = FALSE,
  unlocked_formula = FALSE,
  ...
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet}

\item{dims}{dims}

\item{calculated_column}{calculatedColumn}

\item{empty_cell_reference}{emptyCellReference}

\item{eval_error}{evalError}

\item{formula}{formula}

\item{formula_range}{formulaRange}

\item{list_data_validation}{listDataValidation}

\item{number_stored_as_text}{numberStoredAsText}

\item{two_digit_text_year}{twoDigitTextYear}

\item{unlocked_formula}{unlockedFormula}

\item{...}{additional arguments}
}
\description{
Ignore error on worksheet
}
