% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_ignore_error}
\alias{wb_add_ignore_error}
\title{Ignore error on worksheet}
\usage{
wb_add_ignore_error(
  wb,
  sheet = current_sheet(),
  dims = "A1",
  calculatedColumn = FALSE,
  emptyCellReference = FALSE,
  evalError = FALSE,
  formula = FALSE,
  formulaRange = FALSE,
  listDataValidation = FALSE,
  numberStoredAsText = FALSE,
  twoDigitTextYear = FALSE,
  unlockedFormula = FALSE
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet}

\item{dims}{dims}

\item{calculatedColumn}{calculatedColumn}

\item{emptyCellReference}{emptyCellReference}

\item{evalError}{evalError}

\item{formula}{formula}

\item{formulaRange}{formulaRange}

\item{listDataValidation}{listDataValidation}

\item{numberStoredAsText}{numberStoredAsText}

\item{twoDigitTextYear}{twoDigitTextYear}

\item{unlockedFormula}{unlockedFormula}
}
\description{
Ignore error on worksheet
}
