% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_ignore_error}
\alias{wb_add_ignore_error}
\title{Ignore error types on worksheet}
\usage{
wb_add_ignore_error(
  wb,
  sheet = current_sheet(),
  dims = "A1",
  calculated_column = FALSE,
  empty_cell_reference = FALSE,
  eval_error = FALSE,
  formula = FALSE,
  formula_range = FALSE,
  list_data_validation = FALSE,
  number_stored_as_text = FALSE,
  two_digit_text_year = FALSE,
  unlocked_formula = FALSE,
  ...
)
}
\arguments{
\item{wb}{A workbook}

\item{sheet}{A sheet name or index.}

\item{dims}{Cell range to ignore the error}

\item{calculated_column}{calculatedColumn}

\item{empty_cell_reference}{emptyCellReference}

\item{eval_error}{evalError}

\item{formula}{formula}

\item{formula_range}{formulaRange}

\item{list_data_validation}{listDataValidation}

\item{number_stored_as_text}{If \code{TRUE}, will not display the error if numbers are stored as text.}

\item{two_digit_text_year}{twoDigitTextYear}

\item{unlocked_formula}{unlockedFormula}

\item{...}{additional arguments}
}
\value{
The \code{wbWorkbook} object, invisibly.
}
\description{
This function allows to hide / ignore certain types of errors shown in a worksheet.
}
