% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{wb_add_mips}
\alias{wb_add_mips}
\title{wb add mips}
\usage{
wb_add_mips(wb, mips = NULL)
}
\arguments{
\item{wb}{a workbook}

\item{mips}{a mips string obtained from \code{\link[=wb_get_mips]{wb_get_mips()}} or a global option "openxlsx2.mips_xml_string"}
}
\value{
the workbook invisible
}
\description{
wb add mips
}
