% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_add_mschart}
\alias{wb_add_mschart}
\title{dummy function to add a chart to an existing workbook
currently only a barplot is possible}
\usage{
wb_add_mschart(wb, sheet = next_sheet(), dims = "B2:H8", graph)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{the sheet on which the graph will appear}

\item{dims}{the dimensions where the sheet will appear}

\item{graph}{mschart graph}
}
\description{
dummy function to add a chart to an existing workbook
currently only a barplot is possible
}
\examples{
if (requireNamespace("mschart")) {
require(mschart)
wb <- wb_workbook()

### Scatter
scatter <- ms_scatterchart(data = iris, x = "Sepal.Length", y = "Sepal.Width", group = "Species")
scatter <- chart_settings(scatter, scatterstyle = "marker")

wb <- wb_add_mschart(wb, dims = "F4:L20", graph = scatter)
}
}
