% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_named_style}
\alias{wb_add_named_style}
\title{add named style for cell region}
\usage{
wb_add_named_style(
  wb,
  sheet = current_sheet(),
  dims = "A1",
  name = "Normal",
  font_name = NULL,
  font_size = NULL
)
}
\arguments{
\item{wb}{wbWorkbook}

\item{sheet}{sheet}

\item{dims}{dims}

\item{name}{name}

\item{font_name, font_size}{optional else the default of the theme}
}
\value{
The \code{wbWorkbook}, invisibly
}
\description{
add named style for cell region
}
