% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_plot}
\alias{wb_add_plot}
\title{Insert the current plot into a worksheet}
\usage{
wb_add_plot(
  wb,
  sheet,
  width = 6,
  height = 4,
  xy = NULL,
  startRow = 1,
  startCol = 1,
  fileType = "png",
  units = "in",
  dpi = 300
)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{width}{Width of figure. Defaults to 6in.}

\item{height}{Height of figure . Defaults to 4in.}

\item{xy}{Alternate way to specify startRow and startCol.  A vector of length 2 of form (startcol, startRow)}

\item{startRow}{Row coordinate of upper left corner of figure. \code{xy[[2]]} when xy is given.}

\item{startCol}{Column coordinate of upper left corner of figure. \code{xy[[1]]} when xy is given.}

\item{fileType}{File type of image}

\item{units}{Units of width and height. Can be "in", "cm" or "px"}

\item{dpi}{Image resolution}
}
\description{
The current plot is saved to a temporary image file using dev.copy.
This file is then written to the workbook using wb_add_image.
}
\examples{
\dontrun{
## Create a new workbook
wb <- wb_workbook()

## Add a worksheet
wb$addWorksheet("Sheet 1", gridLines = FALSE)

## create plot objects
require(ggplot2)
p1 <- qplot(mpg,
  data = mtcars, geom = "density",
  fill = as.factor(gear), alpha = I(.5), main = "Distribution of Gas Mileage"
)
p2 <- qplot(age, circumference,
  data = Orange, geom = c("point", "line"), colour = Tree
)

## Insert currently displayed plot to sheet 1, row 1, column 1
print(p1) # plot needs to be showing
wb_add_plot(wb, 1, width = 5, height = 3.5, fileType = "png", units = "in")

## Insert plot 2
print(p2)
wb_add_plot(wb, 1, xy = c("J", 2), width = 16, height = 10, fileType = "png", units = "cm")

## Save workbook
wb_save(wb, "wb_add_plotExample.xlsx", overwrite = TRUE)
}
}
\seealso{
\code{\link[=wb_add_image]{wb_add_image()}}
}
