% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_sparklines}
\alias{wb_add_sparklines}
\title{clone sheets style}
\usage{
wb_add_sparklines(wb, sheet = current_sheet(), sparklines)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet to add the sparklines to}

\item{sparklines}{sparklines object}
}
\description{
clone sheets style
}
