% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_style}
\alias{wb_add_style}
\title{Add a style to a set of cells}
\usage{
wb_add_style(wb, sheet, style, rows, cols, stack = FALSE, gridExpand = FALSE)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}

\item{sheet}{A worksheet to apply the style to.}

\item{style}{A \link{wbStyle} object}

\item{rows, cols}{Rows and columns to apply style to.}

\item{stack}{If \code{TRUE} the new style is merged with any existing cell styles.
If \code{FALSE}, any existing style is replaced by the new style.}

\item{gridExpand}{If \code{TRUE}, style will be applied to all combinations of
\code{rows} and \code{cols.}}
}
\description{
Function adds a style to a specified set of cells.
}
\examples{
## See package vignette for more examples.

## Create a new workbook
wb <- wb_workbook("My name here")

## Add a worksheets
wb$addWorksheet("Expenditure", showGridLines = FALSE)

## write data to worksheet 1
writeData(wb, sheet = 1, USPersonalExpenditure, rowNames = TRUE)

## create and add a style to the column headers
headerStyle <- wbStyle$new(
  fontSize     = 14,
  fontColour   = "#FFFFFF",
  halign       = "center",
  fgFill       = "#4F81BD",
  border       = c("top", "bottom"),
  borderColour = "#4F81BD"
)

wb$addStyle(1, headerStyle, rows = 1, cols = 1:6, gridExpand = TRUE)

## style for body
bodyStyle <- wbStyle$new(border = c("top", "bottom"), borderColour = "#4F81BD")
wb$addStyle(1, bodyStyle, rows = 2:6, cols = 1:6, gridExpand = TRUE)
# set column widht for row names and column
setColWidths(wb, 1, cols = 1, widths = 21)
\dontrun{
wb_save(wb, "addStyleExample.xlsx", overwrite = TRUE)
}
}
\seealso{
\code{\link[=wb_style]{wb_style()}}

Other workbook wrappers: 
\code{\link{setColWidths}()},
\code{\link{setLastModifiedBy}()},
\code{\link{wb_add_worksheet}()},
\code{\link{wb_clone_worksheet}()},
\code{\link{wb_creators}},
\code{\link{wb_freeze_pane}()},
\code{\link{wb_get_base_font}()},
\code{\link{wb_save}()},
\code{\link{wb_set_row_heights}()},
\code{\link{wb_workbook}()},
\code{\link{workbook_grouping}},
\code{\link{ws_cell_merge}}
}
\concept{workbook wrappers}
