% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_add_style}
\alias{wb_add_style}
\title{add style to workbook}
\usage{
wb_add_style(wb, style = NULL, style_name = NULL)
}
\arguments{
\item{wb}{workbook}

\item{style}{style xml character}

\item{style_name}{style name used optional argument}
}
\description{
wb wrapper to add style to workbook
}
\examples{
yellow_f <- wb_colour(hex = "FF9C6500")
yellow_b <- wb_colour(hex = "FFFFEB9C")

yellow <- create_dxfs_style(font_color = yellow_f, bgFill = yellow_b)
wb <- wb_workbook() \%>\% wb_add_style(yellow)
}
\seealso{
\code{\link[=create_border]{create_border()}}, \code{\link[=create_cell_style]{create_cell_style()}}, \code{\link[=create_dxfs_style]{create_dxfs_style()}}, \code{\link[=create_fill]{create_fill()}}, \code{\link[=create_font]{create_font()}}, \code{\link[=create_numfmt]{create_numfmt()}}
}
