% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_base_colors}
\alias{wb_base_colors}
\alias{wb_set_base_colors}
\alias{wb_get_base_colors}
\alias{wb_set_base_colours}
\alias{wb_get_base_colours}
\title{Set the default colors in a workbook}
\usage{
wb_set_base_colors(wb, theme = "Office", ...)

wb_get_base_colors(wb)
}
\arguments{
\item{wb}{A workbook object}

\item{theme}{a predefined color theme}

\item{...}{optional parameters}
}
\description{
Modify / get the default colors of the workbook.
}
