% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_clean_sheet}
\alias{wb_clean_sheet}
\title{Remove all values in a worksheet}
\usage{
wb_clean_sheet(
  wb,
  sheet = current_sheet(),
  dims = NULL,
  numbers = TRUE,
  characters = TRUE,
  styles = TRUE,
  merged_cells = TRUE
)
}
\arguments{
\item{wb}{A Workbook object}

\item{sheet}{sheet to clean}

\item{dims}{spreadsheet dimensions (optional)}

\item{numbers}{remove all numbers}

\item{characters}{remove all characters}

\item{styles}{remove all styles}

\item{merged_cells}{remove all merged_cells}
}
\value{
A Workbook object
}
\description{
Remove content of a worksheet completely, or a region if specifying \code{dims}.
}
