% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_clone_sheet_style}
\alias{wb_clone_sheet_style}
\title{Apply styling from a sheet to another within a workbook}
\usage{
wb_clone_sheet_style(wb, from = current_sheet(), to)
}
\arguments{
\item{wb}{A workbook}

\item{from}{sheet we select the style from}

\item{to}{sheet to apply the style to}
}
\description{
This function can be used to apply styling from a cell range, and apply it
to another cell range.
}
