% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_clone_sheet_style}
\alias{wb_clone_sheet_style}
\title{clone sheets style}
\usage{
wb_clone_sheet_style(wb, from = current_sheet(), to)
}
\arguments{
\item{wb}{workbook}

\item{from}{sheet we select the style from}

\item{to}{sheet we apply the style from}
}
\description{
clone sheets style
}
\seealso{
Other styles: 
\code{\link{wb_add_border}()},
\code{\link{wb_add_cell_style}()},
\code{\link{wb_add_fill}()},
\code{\link{wb_add_font}()},
\code{\link{wb_add_numfmt}()}
}
\concept{styles}
