% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_clone_worksheet}
\alias{wb_clone_worksheet}
\title{Clone a worksheet to a workbook}
\usage{
wb_clone_worksheet(wb, old, new)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}

\item{old}{Name of existing worksheet to copy}

\item{new}{Name of New worksheet to create}
}
\value{
The \code{wb} object
}
\description{
Clone a worksheet to a Workbook object
}
\examples{
# Create a new workbook
wb <- wb_workbook("Fred")

# Add worksheets
wb$addWorksheet("Sheet 1")
wb$cloneWorksheet("Sheet 1", "Sheet 2")

# Save workbook
\dontrun{
wb_save(wb, "cloneWorksheetExample.xlsx", overwrite = TRUE)
}
}
