% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_clone_worksheet}
\alias{wb_clone_worksheet}
\title{Clone a worksheet to a workbook}
\usage{
wb_clone_worksheet(wb, old, new)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet}

\item{old}{The name of the existing worksheet to be cloned.}

\item{new}{A name for the new worksheet}
}
\value{
The \code{wb} object
}
\description{
Clone a worksheet to a Workbook object
}
\examples{
## Create a new workbook
wb <- wb_workbook("Fred")

## Add 3 worksheets
wb$addWorksheet("Sheet 1")
wb$cloneWorksheet("Sheet 1", "Sheet 2")

## Save workbook
\dontrun{
wb_save(wb, "cloneWorksheetExample.xlsx", overwrite = TRUE)
}
}
