% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-color.R
\name{wb_color}
\alias{wb_color}
\alias{wb_colour}
\title{Helper to create a color}
\usage{
wb_color(
  name = NULL,
  auto = NULL,
  indexed = NULL,
  hex = NULL,
  theme = NULL,
  tint = NULL
)
}
\arguments{
\item{name}{A name of a color known to R either as name or RGB/ARGB value.}

\item{auto}{A boolean.}

\item{indexed}{An indexed color value. This color has to be provided by the workbook.}

\item{hex}{A rgb color either a ARGB hex value or RGB hex value With or without leading "#".}

\item{theme}{A zero based index referencing a value in the theme.}

\item{tint}{A tint value applied. Range from -1 (dark) to 1 (light).}
}
\value{
a \code{wbColour} object
}
\description{
Creates a \code{wbColour} object.
}
\seealso{
\code{\link[=wb_get_base_colors]{wb_get_base_colors()}} \code{\link[grDevices:colors]{grDevices::colors()}}
}
