% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_comment}
\alias{wb_comment}
\alias{wb_add_comment}
\alias{wb_remove_comment}
\title{Add / remove comment in worksheet}
\usage{
wb_add_comment(wb, sheet = current_sheet(), dims = "A1", comment, ...)

wb_remove_comment(wb, sheet = current_sheet(), dims = "A1", ...)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A worksheet of the workbook}

\item{dims}{Row and column as spreadsheet dimension, e.g. "A1"}

\item{comment}{A comment to apply to the worksheet}

\item{...}{additional arguments}
}
\value{
The \code{wbWorkbook} object
}
\description{
The comment functions (add and remove) allow the modification of comments.
In newer spreadsheet software they are called notes, while they are called
comments in openxml. Modification of what newer spreadsheet software now calls
comment is possible via \code{\link[=wb_add_thread]{wb_add_thread()}}.

\code{comment} can be created with \code{\link[=create_comment]{create_comment()}} to add styling.
}
\examples{
wb <- wb_workbook()
wb$add_worksheet("Sheet 1")
# Add comment without styling
wb$add_comment(dims = "A1", "Top-left corner of the spreadsheet.")
# Remove comment
wb$remove_comment(sheet = "Sheet 1", dims = "A1")
# add a comment without author
c1 <- create_comment(text = "this is a comment", author = "")
wb$add_comment(dims = "B10", comment = c1)

# Write another comment with author information
c2 <- create_comment(text = "this is another comment", author = "Marco Polo")
wb$add_comment(sheet = 1, dims = "C10", comment = c2)
}
\seealso{
\code{\link[=create_comment]{create_comment()}}, \code{\link[=wb_add_thread]{wb_add_thread()}}
}
