% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_data}
\alias{wb_data}
\title{provide wb_data object as mschart input}
\usage{
wb_data(wb, sheet = current_sheet(), dims = "A1")
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{a sheet in the workbook either name or index}

\item{dims}{the dimensions}
}
\description{
provide wb_data object as mschart input
}
\examples{
 wb <- wb_workbook() \%>\%
   wb_add_worksheet() \%>\%
   wb_add_data(x = mtcars, dims = "B2")

 wb_data(wb, 1, dims = "B2:E6")
}
