% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-named-regions.R
\name{wb_get_named_regions}
\alias{wb_get_named_regions}
\title{Get create or remove named regions}
\usage{
wb_get_named_regions(x, tables = FALSE)
}
\arguments{
\item{x}{An xlsx file or \code{wbWorkbook} object}

\item{tables}{add tables too if \code{TRUE}}
}
\description{
Return a vector of named regions in a xlsx file or \code{wbWorkbook} object
}
