% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_open}
\alias{wb_open}
\title{Open a workbook in a spreadsheet software}
\usage{
wb_open(wb)
}
\arguments{
\item{wb}{a workbook}
}
\description{
Open a workbook in a spreadsheet software
}
