% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_open}
\alias{wb_open}
\title{Preview a workbook in a spreadsheet software}
\usage{
wb_open(wb, interactive = NA, flush = FALSE)
}
\arguments{
\item{wb}{a \link{wbWorkbook} object}

\item{interactive}{If \code{FALSE} will throw a warning and not open the path.
This can be manually set to \code{TRUE}, otherwise when \code{NA} (default) uses
the value returned from \code{\link[base:interactive]{base::interactive()}}}

\item{flush}{if the flush option should be used}
}
\description{
You can also use the shorter \code{wb$open()} as a replacement.
To open xlsx files, see \code{\link[=xl_open]{xl_open()}}.
}
