% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_open}
\alias{wb_open}
\title{Preview a workbook in a spreadsheet software}
\usage{
wb_open(wb)
}
\arguments{
\item{wb}{a \link{wbWorkbook} object}
}
\description{
You can also use the shorter \code{wb$open()} as a replacement.
To open xlsx files, see \code{\link[=xl_open]{xl_open()}}.
}
