% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorkbook.R
\name{wb_read}
\alias{wb_read}
\title{Read from an Excel file or Workbook object}
\usage{
wb_read(
  xlsx_file,
  sheet = 1,
  start_row = 1,
  start_col = NULL,
  row_names = FALSE,
  col_names = TRUE,
  skip_empty_rows = FALSE,
  skip_empty_cols = FALSE,
  rows = NULL,
  cols = NULL,
  detect_dates = TRUE,
  named_region,
  na.strings = "NA",
  na.numbers = NA,
  ...
)
}
\arguments{
\item{xlsx_file}{An xlsx file, Workbook object or URL to xlsx file.}

\item{sheet}{The name or index of the sheet to read data from.}

\item{start_row}{first row to begin looking for data.}

\item{start_col}{first column to begin looking for data.}

\item{row_names}{If \code{TRUE}, first column of data will be used as row names.}

\item{col_names}{If \code{TRUE}, the first row of data will be used as column names.}

\item{skip_empty_rows}{If \code{TRUE}, empty rows are skipped else empty rows after the first row containing data
will return a row of NAs.}

\item{skip_empty_cols}{If \code{TRUE}, empty columns are skipped.}

\item{rows}{A numeric vector specifying which rows in the Excel file to read.
If NULL, all rows are read.}

\item{cols}{A numeric vector specifying which columns in the Excel file to read.
If NULL, all columns are read.}

\item{detect_dates}{If \code{TRUE}, attempt to recognize dates and perform conversion.}

\item{named_region}{A named region in the Workbook. If not NULL startRow, rows and cols parameters are ignored.}

\item{na.strings}{A character vector of strings which are to be interpreted as NA. Blank cells will be returned as NA.}

\item{na.numbers}{A numeric vector of digits which are to be interpreted as NA. Blank cells will be returned as NA.}

\item{...}{additional arguments passed to \code{wb_to_df()}}
}
\value{
A data.frame
}
\description{
Read data from an Excel file or Workbook object into a data.frame
}
\details{
Creates a data.frame of all data in worksheet.
}
\examples{
xlsxFile <- system.file("extdata", "openxlsx2_example.xlsx", package = "openxlsx2")
df1 <- wb_read(xlsxFile = xlsxFile, sheet = 1)

xlsxFile <- system.file("extdata", "openxlsx2_example.xlsx", package = "openxlsx2")
df1 <- wb_read(xlsxFile = xlsxFile, sheet = 1, rows = c(1, 3, 5), cols = 1:3)
}
\seealso{
\code{\link[=wb_get_named_regions]{wb_get_named_regions()}} \code{\link[=wb_to_df]{wb_to_df()}} \code{\link[=read_xlsx]{read_xlsx()}}
}
