% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_save}
\alias{wb_save}
\title{Save Workbook to file}
\usage{
wb_save(wb, path = NULL, overwrite = TRUE)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object to write to file}

\item{path}{A path to save the workbook to}

\item{overwrite}{If \code{FALSE}, will not overwrite when \code{path} exists}
}
\description{
Save Workbook to file
}
\examples{
## Create a new workbook and add a worksheet
wb <- wb_workbook("Creator of workbook")
wb$addWorksheet(sheetName = "My first worksheet")

## Save workbook to working directory
\dontrun{
wb_save(wb, path = temp_xlsx(), overwrite = TRUE)
}
}
\seealso{
\code{\link[=wb_workbook]{wb_workbook()}}, \code{\link[=wb_add_worksheet]{wb_add_worksheet()}}, \code{\link[=loadWorkbook]{loadWorkbook()}}, \code{\link[=writeData]{writeData()}}.
\code{\link[=writeDataTable]{writeDataTable()}}
}
