% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_bookview}
\alias{wb_set_bookview}
\title{Set the workbook position, size and filter}
\usage{
wb_set_bookview(
  wb,
  active_tab = NULL,
  auto_filter_date_grouping = NULL,
  first_sheet = NULL,
  minimized = NULL,
  show_horizontal_scroll = NULL,
  show_sheet_tabs = NULL,
  show_vertical_scroll = NULL,
  tab_ratio = NULL,
  visibility = NULL,
  window_height = NULL,
  window_width = NULL,
  x_window = NULL,
  y_window = NULL,
  ...
)
}
\arguments{
\item{wb}{A \link{wbWorkbook} object}

\item{active_tab}{activeTab}

\item{auto_filter_date_grouping}{autoFilterDateGrouping}

\item{first_sheet}{The first sheet to be displayed}

\item{minimized}{minimized}

\item{show_horizontal_scroll}{showHorizontalScroll}

\item{show_sheet_tabs}{showSheetTabs}

\item{show_vertical_scroll}{showVerticalScroll}

\item{tab_ratio}{tabRatio}

\item{visibility}{visibility}

\item{window_height}{windowHeight}

\item{window_width}{windowWidth}

\item{x_window}{xWindow}

\item{y_window}{yWindow}

\item{...}{additional arguments}
}
\value{
The Workbook object
}
\description{
Set the workbook position, size and filter
}
