% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_last_modified_by}
\alias{wb_set_last_modified_by}
\title{Add another author to the meta data of the file.}
\usage{
wb_set_last_modified_by(wb, name, ...)
}
\arguments{
\item{wb}{A workbook object}

\item{name}{A string object with the name of the LastModifiedBy-User}

\item{...}{additional arguments}
}
\description{
Just a wrapper of \code{wb$set_last_modified_by()}
}
\examples{
wb <- wb_workbook()
wb_set_last_modified_by(wb, "test")
}
