% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_last_modified_by}
\alias{wb_set_last_modified_by}
\title{Add another author to the meta data of the file.}
\usage{
wb_set_last_modified_by(wb, LastModifiedBy)
}
\arguments{
\item{wb}{A workbook object}

\item{LastModifiedBy}{A string object with the name of the LastModifiedBy-User}
}
\description{
Just a wrapper of wb$set_last_modified_by()
}
\examples{
wb <- wb_workbook()
wb_set_last_modified_by(wb, "test")
}
\seealso{
Other workbook wrappers: 
\code{\link{wb_add_data_table}()},
\code{\link{wb_add_data}()},
\code{\link{wb_add_formula}()},
\code{\link{wb_add_worksheet}()},
\code{\link{wb_clone_worksheet}()},
\code{\link{wb_creators}},
\code{\link{wb_freeze_pane}()},
\code{\link{wb_get_base_font}()},
\code{\link{wb_save}()},
\code{\link{wb_set_col_widths}()},
\code{\link{wb_set_row_heights}()},
\code{\link{wb_workbook}()},
\code{\link{workbook_grouping}},
\code{\link{ws_cell_merge}}
}
\concept{workbook wrappers}
