% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_last_modified_by}
\alias{wb_set_last_modified_by}
\title{Modify author in the metadata of a workbook}
\usage{
wb_set_last_modified_by(wb, name, ...)
}
\arguments{
\item{wb}{A workbook object}

\item{name}{A string object with the name of the LastModifiedBy-User}

\item{...}{additional arguments}
}
\description{
Just a wrapper of \code{wb$set_last_modified_by()}
}
\examples{
wb <- wb_workbook()
wb_set_last_modified_by(wb, "test")
}
\seealso{
Other workbook wrappers: 
\code{\link{base_font-wb}},
\code{\link{col_widths-wb}},
\code{\link{creators-wb}},
\code{\link{grouping-wb}},
\code{\link{row_heights-wb}},
\code{\link{wb_add_chartsheet}()},
\code{\link{wb_add_data_table}()},
\code{\link{wb_add_data}()},
\code{\link{wb_add_formula}()},
\code{\link{wb_add_pivot_table}()},
\code{\link{wb_add_worksheet}()},
\code{\link{wb_clone_worksheet}()},
\code{\link{wb_copy_cells}()},
\code{\link{wb_freeze_pane}()},
\code{\link{wb_merge_cells}()},
\code{\link{wb_save}()},
\code{\link{wb_workbook}()}
}
\concept{workbook wrappers}
