% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_properties}
\alias{wb_set_properties}
\title{Modify workbook properties}
\usage{
wb_set_properties(wb, title = NULL, subject = NULL, category = NULL)
}
\arguments{
\item{title, subject, category}{Workbook property, a string.}
}
\description{
Modify workbook properties
}
