% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_properties}
\alias{wb_set_properties}
\title{Modify workbook properties}
\usage{
wb_set_properties(wb, title = NULL, subject = NULL, category = NULL)
}
\arguments{
\item{title, subject, category}{Workbook property, a string.}
}
\value{
A wbWorkbook object, invisibly.
}
\description{
This function is useful for workbooks that are loeaded
}
\examples{
file <- system.file("extdata", "openxlsx2_example.xlsx", package = "openxlsx2")
wb <- wb_load(file)
wb$title

# Add a title to properties
wb$set_properties(title = "my title")

wb$title

}
