% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_set_sheetview}
\alias{wb_set_sheetview}
\title{add sheetview}
\usage{
wb_set_sheetview(
  wb,
  sheet = current_sheet(),
  colorId = NULL,
  defaultGridColor = NULL,
  rightToLeft = NULL,
  showFormulas = NULL,
  showGridLines = NULL,
  showOutlineSymbols = NULL,
  showRowColHeaders = NULL,
  showRuler = NULL,
  showWhiteSpace = NULL,
  showZeros = NULL,
  tabSelected = NULL,
  topLeftCell = NULL,
  view = NULL,
  windowProtection = NULL,
  workbookViewId = NULL,
  zoomScale = NULL,
  zoomScaleNormal = NULL,
  zoomScalePageLayoutView = NULL,
  zoomScaleSheetLayoutView = NULL
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet}

\item{colorId}{colorId}

\item{defaultGridColor}{defaultGridColor}

\item{rightToLeft}{rightToLeft}

\item{showFormulas}{showFormulas}

\item{showGridLines}{showGridLines}

\item{showOutlineSymbols}{showOutlineSymbols}

\item{showRowColHeaders}{showRowColHeaders}

\item{showRuler}{showRuler}

\item{showWhiteSpace}{showWhiteSpace}

\item{showZeros}{showZeros}

\item{tabSelected}{tabSelected}

\item{topLeftCell}{topLeftCell}

\item{view}{view}

\item{windowProtection}{windowProtection}

\item{workbookViewId}{workbookViewId}

\item{zoomScale}{zoomScale}

\item{zoomScaleNormal}{zoomScaleNormal}

\item{zoomScalePageLayoutView}{zoomScalePageLayoutView}

\item{zoomScaleSheetLayoutView}{zoomScaleSheetLayoutView}
}
\value{
The \code{wbWorksheetObject}, invisibly
}
\description{
add sheetview
}
