% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_update_table}
\alias{wb_update_table}
\title{update a data_table}
\usage{
wb_update_table(wb, sheet = current_sheet(), dims = "A1", tabname)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{a worksheet}

\item{dims}{cell used as start}

\item{tabname}{a tablename}
}
\description{
update a data_table
}
\details{
Be aware that this function does not alter any filter. Excluding or adding rows does not make rows appear nor will it hide them.
}
\examples{
wb <- wb_workbook()$add_worksheet()$add_data_table(x = mtcars)
wb$update_table(tabname = "Table1", dims = "A1:J4")
}
