% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_workbook}
\alias{wb_workbook}
\title{Create a new Workbook object}
\usage{
wb_workbook(creator = NULL, title = NULL, subject = NULL, category = NULL)
}
\arguments{
\item{creator}{Creator of the workbook (your name). Defaults to login username}

\item{title}{Workbook properties title}

\item{subject}{Workbook properties subject}

\item{category}{Workbook properties category}
}
\value{
Workbook object
}
\description{
Create a new Workbook object
}
\examples{
## Create a new workbook
wb <- wb_workbook()

## Save workbook to working directory
\dontrun{
wb_save(wb, path = temp_xlsx(), overwrite = TRUE)
}

## Set Workbook properties
wb <- wb_workbook(
  creator  = "Me",
  title    = "Expense Report",
  subject  = "Expense Report - 2022 Q1",
  category = "sales"
)
}
\seealso{
\code{\link[=loadWorkbook]{loadWorkbook()}}

\code{\link[=wb_save]{wb_save()}}
}
