% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook-wrappers.R
\name{wb_workbook}
\alias{wb_workbook}
\title{Create a new Workbook object}
\usage{
wb_workbook(creator = NULL, title = NULL, subject = NULL, category = NULL)
}
\arguments{
\item{creator}{Creator(s) of the workbook (see detaisl)}

\item{title, subject, category}{Title, subject, and category of the workmust.
These must be \code{NULL} or single length \code{character} vectors}
}
\value{
A \code{wbWorkbook} object
}
\description{
Create a new Workbook object
}
\details{
When \code{creator} is not set, a search is made for a non-\code{NULL} option
in \code{openxlsx.creator}.  If none is found, the environment variables
\code{USERNAME} and \code{USER} are searched, in that order.
}
\examples{
## Create a new workbook
wb <- wb_workbook()

## Save workbook to working directory
\dontrun{
wb_save(wb, path = temp_xlsx(), overwrite = TRUE)
}

## Set Workbook properties
wb <- wb_workbook(
  creator  = "Me",
  title    = "Expense Report",
  subject  = "Expense Report - 2022 Q1",
  category = "sales"
)
}
\seealso{
\code{\link[=loadWorkbook]{loadWorkbook()}}, \code{\link[=wb_save]{wb_save()}}
}
