% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-chart-sheet.R
\name{wbChartSheet}
\alias{wbChartSheet}
\title{R6 class for a Workbook Chart Sheet}
\value{
A character vector of xml
}
\description{
R6 class for a Workbook Chart Sheet

R6 class for a Workbook Chart Sheet
}
\details{
A chart sheet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sheetPr}}{Sheet something?}

\item{\code{sheetViews}}{Something}

\item{\code{pageMargins}}{page margins}

\item{\code{drawing}}{drawing}

\item{\code{hyperlinks}}{hyperlinks}

\item{\code{relships}}{relships}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wbChartSheet-new}{\code{wbChartSheet$new()}}
\item \href{#method-wbChartSheet-get_prior_sheet_data}{\code{wbChartSheet$get_prior_sheet_data()}}
\item \href{#method-wbChartSheet-set_sheetview}{\code{wbChartSheet$set_sheetview()}}
\item \href{#method-wbChartSheet-clone}{\code{wbChartSheet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbChartSheet-new"></a>}}
\if{latex}{\out{\hypertarget{method-wbChartSheet-new}{}}}
\subsection{Method \code{new()}}{
Create a new workbook chart sheet object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbChartSheet$new(tabColor = tabColor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tabColor}}{\code{character} a tab color to set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbChartSheet} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbChartSheet-get_prior_sheet_data"></a>}}
\if{latex}{\out{\hypertarget{method-wbChartSheet-get_prior_sheet_data}{}}}
\subsection{Method \code{get_prior_sheet_data()}}{
get (prior) sheet data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbChartSheet$get_prior_sheet_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbChartSheet-set_sheetview"></a>}}
\if{latex}{\out{\hypertarget{method-wbChartSheet-set_sheetview}{}}}
\subsection{Method \code{set_sheetview()}}{
add sheetview
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbChartSheet$set_sheetview(
  colorId = NULL,
  defaultGridColor = NULL,
  rightToLeft = NULL,
  showFormulas = NULL,
  showGridLines = NULL,
  showOutlineSymbols = NULL,
  showRowColHeaders = NULL,
  showRuler = NULL,
  showWhiteSpace = NULL,
  showZeros = NULL,
  tabSelected = NULL,
  topLeftCell = NULL,
  view = NULL,
  windowProtection = NULL,
  workbookViewId = NULL,
  zoomScale = NULL,
  zoomScaleNormal = NULL,
  zoomScalePageLayoutView = NULL,
  zoomScaleSheetLayoutView = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colorId}}{colorId}

\item{\code{defaultGridColor}}{defaultGridColor}

\item{\code{rightToLeft}}{rightToLeft}

\item{\code{showFormulas}}{showFormulas}

\item{\code{showGridLines}}{showGridLines}

\item{\code{showOutlineSymbols}}{showOutlineSymbols}

\item{\code{showRowColHeaders}}{showRowColHeaders}

\item{\code{showRuler}}{showRuler}

\item{\code{showWhiteSpace}}{showWhiteSpace}

\item{\code{showZeros}}{showZeros}

\item{\code{tabSelected}}{tabSelected}

\item{\code{topLeftCell}}{topLeftCell}

\item{\code{view}}{view}

\item{\code{windowProtection}}{windowProtection}

\item{\code{workbookViewId}}{workbookViewId}

\item{\code{zoomScale}}{zoomScale}

\item{\code{zoomScaleNormal}}{zoomScaleNormal}

\item{\code{zoomScalePageLayoutView}}{zoomScalePageLayoutView}

\item{\code{zoomScaleSheetLayoutView}}{zoomScaleSheetLayoutView}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorksheetObject}, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbChartSheet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wbChartSheet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbChartSheet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
