% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sheet-data.R
\name{wbSheetData}
\alias{wbSheetData}
\alias{wb_sheet_data}
\title{R6 class for a Workbook Hyperlink}
\usage{
wb_sheet_data()
}
\description{
R6 class for a Workbook Hyperlink

R6 class for a Workbook Hyperlink
}
\details{
A hyperlink
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{row_attr}}{row_attr}

\item{\code{cc}}{cc}

\item{\code{cc_out}}{cc_out}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wbSheetData-new}{\code{wbSheetData$new()}}
\item \href{#method-wbSheetData-clone}{\code{wbSheetData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbSheetData-new"></a>}}
\if{latex}{\out{\hypertarget{method-wbSheetData-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbSheetData} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbSheetData$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{wbSheetData} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wbSheetData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wbSheetData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbSheetData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
