% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-workbook.R
\name{wbWorkbook}
\alias{wbWorkbook}
\title{R6 class for a Workbook}
\value{
The \code{wbWorkbook} object

The sheet tables.  \code{character()} if empty

The \code{wbWorkbook} object

The \code{wbWorkbook} object

The \code{wbWorkbook} object

The \code{wbWorkbook} object

The \code{wbWorkbook} object

The \code{wbWorkbook} object

Returns sheet visibility

The \code{wbWorkbook} object
Deep cloning method for workbooks.  This method also accesses
\verb{$clone(deep = TRUE)} methods for \code{R6} fields.
}
\description{
R6 class for a Workbook

R6 class for a Workbook
}
\details{
A Workbook
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sheet_names}}{sheet_names}

\item{\code{calcChain}}{calcChain}

\item{\code{apps}}{apps}

\item{\code{charts}}{charts}

\item{\code{isChartSheet}}{isChartSheet}

\item{\code{connections}}{connections}

\item{\code{Content_Types}}{Content_Types}

\item{\code{app}}{app}

\item{\code{core}}{core}

\item{\code{drawings}}{drawings}

\item{\code{drawings_rels}}{drawings_rels}

\item{\code{drawings_vml}}{drawings_vml}

\item{\code{embeddings}}{embeddings}

\item{\code{externalLinks}}{externalLinks}

\item{\code{externalLinksRels}}{externalLinksRels}

\item{\code{headFoot}}{headFoot}

\item{\code{media}}{media}

\item{\code{persons}}{persons}

\item{\code{pivotTables}}{pivotTables}

\item{\code{pivotTables.xml.rels}}{pivotTables.xml.rels}

\item{\code{pivotDefinitions}}{pivotDefinitions}

\item{\code{pivotRecords}}{pivotRecords}

\item{\code{pivotDefinitionsRels}}{pivotDefinitionsRels}

\item{\code{queryTables}}{queryTables}

\item{\code{rowHeights}}{rowHeights}

\item{\code{slicers}}{slicers}

\item{\code{slicerCaches}}{slicerCaches}

\item{\code{sharedStrings}}{sharedStrings}

\item{\code{styles_mgr}}{styles_mgr}

\item{\code{styles_xml}}{styles_xml}

\item{\code{tables}}{tables}

\item{\code{tables.xml.rels}}{tables.xml.rels}

\item{\code{theme}}{theme}

\item{\code{vbaProject}}{vbaProject}

\item{\code{vml}}{vml}

\item{\code{vml_rels}}{vml_rels}

\item{\code{comments}}{comments}

\item{\code{threadComments}}{threadComments}

\item{\code{workbook}}{workbook}

\item{\code{workbook.xml.rels}}{workbook.xml.rels}

\item{\code{worksheets}}{worksheets}

\item{\code{worksheets_rels}}{worksheets_rels}

\item{\code{sheetOrder}}{sheetOrder}

\item{\code{path}}{path}

\item{\code{creator}}{A character vector of creators}

\item{\code{title}}{title}

\item{\code{subject}}{subject}

\item{\code{category}}{category}

\item{\code{datetimeCreated}}{The datetime (as \code{POSIXt}) the workbook is
created.  Defaults to the current \code{Sys.time()} when the workbook object
is created, not when the Excel files are saved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{wbWorkbook$new()}}
\item \href{#method-append}{\code{wbWorkbook$append()}}
\item \href{#method-append_sheets}{\code{wbWorkbook$append_sheets()}}
\item \href{#method-add_worksheet}{\code{wbWorkbook$add_worksheet()}}
\item \href{#method-clone_worksheet}{\code{wbWorkbook$clone_worksheet()}}
\item \href{#method-addChartSheet}{\code{wbWorkbook$addChartSheet()}}
\item \href{#method-save}{\code{wbWorkbook$save()}}
\item \href{#method-buildTable}{\code{wbWorkbook$buildTable()}}
\item \href{#method-createFontNode}{\code{wbWorkbook$createFontNode()}}
\item \href{#method-get_base_font}{\code{wbWorkbook$get_base_font()}}
\item \href{#method-set_base_font}{\code{wbWorkbook$set_base_font()}}
\item \href{#method-setSheetName}{\code{wbWorkbook$setSheetName()}}
\item \href{#method-set_row_heights}{\code{wbWorkbook$set_row_heights()}}
\item \href{#method-remove_row_heights}{\code{wbWorkbook$remove_row_heights()}}
\item \href{#method-createCols}{\code{wbWorkbook$createCols()}}
\item \href{#method-group_cols}{\code{wbWorkbook$group_cols()}}
\item \href{#method-remove_col_widths}{\code{wbWorkbook$remove_col_widths()}}
\item \href{#method-set_col_widths}{\code{wbWorkbook$set_col_widths()}}
\item \href{#method-wb_group_rows}{\code{wbWorkbook$wb_group_rows()}}
\item \href{#method-remove_worksheet}{\code{wbWorkbook$remove_worksheet()}}
\item \href{#method-data_validation}{\code{wbWorkbook$data_validation()}}
\item \href{#method-data_validation_list}{\code{wbWorkbook$data_validation_list()}}
\item \href{#method-conditional_formatting}{\code{wbWorkbook$conditional_formatting()}}
\item \href{#method-merge_cells}{\code{wbWorkbook$merge_cells()}}
\item \href{#method-unmerge_cells}{\code{wbWorkbook$unmerge_cells()}}
\item \href{#method-freeze_pane}{\code{wbWorkbook$freeze_pane()}}
\item \href{#method-add_image}{\code{wbWorkbook$add_image()}}
\item \href{#method-add_plot}{\code{wbWorkbook$add_plot()}}
\item \href{#method-print}{\code{wbWorkbook$print()}}
\item \href{#method-protect}{\code{wbWorkbook$protect()}}
\item \href{#method-set_creators}{\code{wbWorkbook$set_creators()}}
\item \href{#method-add_creators}{\code{wbWorkbook$add_creators()}}
\item \href{#method-remove_creators}{\code{wbWorkbook$remove_creators()}}
\item \href{#method-set_last_modified_by}{\code{wbWorkbook$set_last_modified_by()}}
\item \href{#method-page_setup}{\code{wbWorkbook$page_setup()}}
\item \href{#method-set_header_footer}{\code{wbWorkbook$set_header_footer()}}
\item \href{#method-get_tables}{\code{wbWorkbook$get_tables()}}
\item \href{#method-remove_tables}{\code{wbWorkbook$remove_tables()}}
\item \href{#method-add_filter}{\code{wbWorkbook$add_filter()}}
\item \href{#method-remove_filter}{\code{wbWorkbook$remove_filter()}}
\item \href{#method-grid_lines}{\code{wbWorkbook$grid_lines()}}
\item \href{#method-add_named_region}{\code{wbWorkbook$add_named_region()}}
\item \href{#method-remove_named_region}{\code{wbWorkbook$remove_named_region()}}
\item \href{#method-set_order}{\code{wbWorkbook$set_order()}}
\item \href{#method-get_sheet_visibility}{\code{wbWorkbook$get_sheet_visibility()}}
\item \href{#method-set_sheet_visibility}{\code{wbWorkbook$set_sheet_visibility()}}
\item \href{#method-clone}{\code{wbWorkbook$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{wbWorkbook} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$new(
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  datetimeCreated = Sys.time()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creator}}{character vector of creators.  Duplicated are ignored.}

\item{\code{title}}{title}

\item{\code{subject}}{subject}

\item{\code{category}}{category}

\item{\code{datetimeCreated}}{The datetime (as \code{POSIXt}) the workbook is
created.  Defaults to the current \code{Sys.time()} when the workbook object
is created, not when the Excel files are saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{wbWorkbook} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\if{latex}{\out{\hypertarget{method-append}{}}}
\subsection{Method \code{append()}}{
Append a field. This method is intended for internal use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$append(field, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{A valid field name}

\item{\code{value}}{A value for the field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_sheets"></a>}}
\if{latex}{\out{\hypertarget{method-append_sheets}{}}}
\subsection{Method \code{append_sheets()}}{
Append to \code{self$workbook$sheets} This method is intended for internal use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$append_sheets(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A value for \code{self$workbook$sheets}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_worksheet"></a>}}
\if{latex}{\out{\hypertarget{method-add_worksheet}{}}}
\subsection{Method \code{add_worksheet()}}{
Add worksheet to the \code{wbWorkbook} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$add_worksheet(
  sheet,
  gridLines = TRUE,
  tabColour = NULL,
  zoom = 100,
  header = NULL,
  footer = NULL,
  oddHeader = header,
  oddFooter = footer,
  evenHeader = header,
  evenFooter = footer,
  firstHeader = header,
  firstFooter = footer,
  visible = c("true", "false", "hidden", "visible", "veryhidden"),
  hasDrawing = FALSE,
  paperSize = getOption("openxlsx2.paperSize", default = 9),
  orientation = getOption("openxlsx2.orientation", default = "portrait"),
  hdpi = getOption("openxlsx2.hdpi", default = getOption("openxlsx2.dpi", default =
    300)),
  vdpi = getOption("openxlsx2.vdpi", default = getOption("openxlsx2.dpi", default =
    300))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{gridLines}}{gridLines}

\item{\code{tabColour}}{tabColour}

\item{\code{zoom}}{zoom}

\item{\code{header}}{header}

\item{\code{footer}}{footer}

\item{\code{oddHeader}}{oddHeader}

\item{\code{oddFooter}}{oddFooter}

\item{\code{evenHeader}}{evenHeader}

\item{\code{evenFooter}}{evenFooter}

\item{\code{firstHeader}}{firstHeader}

\item{\code{firstFooter}}{firstFooter}

\item{\code{visible}}{visible}

\item{\code{hasDrawing}}{hasDrawing}

\item{\code{paperSize}}{paperSize}

\item{\code{orientation}}{orientation}

\item{\code{hdpi}}{hdpi}

\item{\code{vdpi}}{vdpi}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone_worksheet"></a>}}
\if{latex}{\out{\hypertarget{method-clone_worksheet}{}}}
\subsection{Method \code{clone_worksheet()}}{
Clone a workbooksheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$clone_worksheet(old, new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old}}{name of worksheet to clone}

\item{\code{new}}{name of new worksheet to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addChartSheet"></a>}}
\if{latex}{\out{\hypertarget{method-addChartSheet}{}}}
\subsection{Method \code{addChartSheet()}}{
Add a chart sheet to the workbook
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$addChartSheet(sheet, tabColour = NULL, zoom = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{tabColour}}{tabColour}

\item{\code{zoom}}{zoom}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
Save the workbook
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$save(path = self$path, overwrite = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to save the workbook to}

\item{\code{overwrite}}{If \code{FALSE}, will not overwrite when \code{path} exists}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildTable"></a>}}
\if{latex}{\out{\hypertarget{method-buildTable}{}}}
\subsection{Method \code{buildTable()}}{
Build table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$buildTable(
  sheet,
  colNames,
  ref,
  showColNames,
  tableStyle,
  tableName,
  withFilter,
  totalsRowCount = 0,
  showFirstColumn = 0,
  showLastColumn = 0,
  showRowStripes = 1,
  showColumnStripes = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{colNames}}{colNames}

\item{\code{ref}}{ref}

\item{\code{showColNames}}{showColNames}

\item{\code{tableStyle}}{tableStyle}

\item{\code{tableName}}{tableName}

\item{\code{withFilter}}{withFilter}

\item{\code{totalsRowCount}}{totalsRowCount}

\item{\code{showFirstColumn}}{showFirstColumn}

\item{\code{showLastColumn}}{showLastColumn}

\item{\code{showRowStripes}}{showRowStripes}

\item{\code{showColumnStripes}}{showColumnStripes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorksheet} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createFontNode"></a>}}
\if{latex}{\out{\hypertarget{method-createFontNode}{}}}
\subsection{Method \code{createFontNode()}}{
Create a font node from a style
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$createFontNode(style)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{style}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The font node as xml?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_base_font"></a>}}
\if{latex}{\out{\hypertarget{method-get_base_font}{}}}
\subsection{Method \code{get_base_font()}}{
Get the base font
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$get_base_font()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of of the font
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_base_font"></a>}}
\if{latex}{\out{\hypertarget{method-set_base_font}{}}}
\subsection{Method \code{set_base_font()}}{
Get the base font
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_base_font(
  fontSize = 11,
  fontColour = "black",
  fontName = "Calibri"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fontSize}}{fontSize}

\item{\code{fontColour}}{fontColour}

\item{\code{fontName}}{fontName}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSheetName"></a>}}
\if{latex}{\out{\hypertarget{method-setSheetName}{}}}
\subsection{Method \code{setSheetName()}}{
Sets a sheet name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$setSheetName(sheet, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{Old sheet name}

\item{\code{name}}{New sheet name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_row_heights"></a>}}
\if{latex}{\out{\hypertarget{method-set_row_heights}{}}}
\subsection{Method \code{set_row_heights()}}{
Sets a row height for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_row_heights(sheet, rows, heights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{rows}}{rows}

\item{\code{heights}}{heights}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_row_heights"></a>}}
\if{latex}{\out{\hypertarget{method-remove_row_heights}{}}}
\subsection{Method \code{remove_row_heights()}}{
Sets a row height for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_row_heights(sheet, rows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{rows}}{rows}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
description
creates column object for worksheet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createCols"></a>}}
\if{latex}{\out{\hypertarget{method-createCols}{}}}
\subsection{Method \code{createCols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$createCols(sheet, n, beg, end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{n}}{n}

\item{\code{beg}}{beg}

\item{\code{end}}{end}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-group_cols"></a>}}
\if{latex}{\out{\hypertarget{method-group_cols}{}}}
\subsection{Method \code{group_cols()}}{
Group cols
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$group_cols(sheet, cols, collapsed, levels = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{cols}}{cols}

\item{\code{collapsed}}{collapsed}

\item{\code{levels}}{levels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_col_widths"></a>}}
\if{latex}{\out{\hypertarget{method-remove_col_widths}{}}}
\subsection{Method \code{remove_col_widths()}}{
Remove row heights from a worksheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_col_widths(sheet, cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{A name or index of a worksheet}

\item{\code{cols}}{Indices of columns to remove custom width (if any) from.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_col_widths"></a>}}
\if{latex}{\out{\hypertarget{method-set_col_widths}{}}}
\subsection{Method \code{set_col_widths()}}{
Group cols
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_col_widths(sheet, cols, widths = 8.43, hidden = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{cols}}{cols}

\item{\code{widths}}{Width of columns}

\item{\code{hidden}}{A logical vector to determine which cols are hidden; values
are repeated across length of \code{cols}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wb_group_rows"></a>}}
\if{latex}{\out{\hypertarget{method-wb_group_rows}{}}}
\subsection{Method \code{wb_group_rows()}}{
Group rows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$wb_group_rows(sheet, rows, collapsed = FALSE, levels = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{rows}}{rows}

\item{\code{collapsed}}{collapsed}

\item{\code{levels}}{levels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_worksheet"></a>}}
\if{latex}{\out{\hypertarget{method-remove_worksheet}{}}}
\subsection{Method \code{remove_worksheet()}}{
Remove a worksheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_worksheet(sheet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{The worksheet to delete}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_validation"></a>}}
\if{latex}{\out{\hypertarget{method-data_validation}{}}}
\subsection{Method \code{data_validation()}}{
Set data validations for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$data_validation(
  sheet,
  startRow,
  endRow,
  startCol,
  endCol,
  type,
  operator,
  value,
  allowBlank,
  showInputMsg,
  showErrorMsg
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{startRow}}{startRow}

\item{\code{endRow}}{endRow}

\item{\code{startCol}}{startCol}

\item{\code{endCol}}{endCol}

\item{\code{type}}{type}

\item{\code{operator}}{operator}

\item{\code{value}}{value}

\item{\code{allowBlank}}{allowBlank}

\item{\code{showInputMsg}}{showInputMsg}

\item{\code{showErrorMsg}}{showErrorMsg}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_validation_list"></a>}}
\if{latex}{\out{\hypertarget{method-data_validation_list}{}}}
\subsection{Method \code{data_validation_list()}}{
Set data validations for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$data_validation_list(
  sheet,
  startRow,
  endRow,
  startCol,
  endCol,
  value,
  allowBlank,
  showInputMsg,
  showErrorMsg
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{startRow}}{startRow}

\item{\code{endRow}}{endRow}

\item{\code{startCol}}{startCol}

\item{\code{endCol}}{endCol}

\item{\code{value}}{value}

\item{\code{allowBlank}}{allowBlank}

\item{\code{showInputMsg}}{showInputMsg}

\item{\code{showErrorMsg}}{showErrorMsg}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conditional_formatting"></a>}}
\if{latex}{\out{\hypertarget{method-conditional_formatting}{}}}
\subsection{Method \code{conditional_formatting()}}{
Set conditional formatting for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$conditional_formatting(
  sheet,
  startRow,
  endRow,
  startCol,
  endCol,
  dxfId,
  formula,
  type,
  values,
  params
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{startRow}}{startRow}

\item{\code{endRow}}{endRow}

\item{\code{startCol}}{startCol}

\item{\code{endCol}}{endCol}

\item{\code{dxfId}}{dxfId}

\item{\code{formula}}{formula}

\item{\code{type}}{type}

\item{\code{values}}{values}

\item{\code{params}}{params}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge_cells"></a>}}
\if{latex}{\out{\hypertarget{method-merge_cells}{}}}
\subsection{Method \code{merge_cells()}}{
Set cell merging for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$merge_cells(sheet, rows = NULL, cols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{rows, cols}}{Row and column specifications.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unmerge_cells"></a>}}
\if{latex}{\out{\hypertarget{method-unmerge_cells}{}}}
\subsection{Method \code{unmerge_cells()}}{
Removes cell merging for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$unmerge_cells(sheet, rows = NULL, cols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{rows, cols}}{Row and column specifications.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-freeze_pane"></a>}}
\if{latex}{\out{\hypertarget{method-freeze_pane}{}}}
\subsection{Method \code{freeze_pane()}}{
Set freeze panes for a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$freeze_pane(
  sheet,
  firstActiveRow = NULL,
  firstActiveCol = NULL,
  firstRow = FALSE,
  firstCol = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{firstActiveRow}}{firstActiveRow}

\item{\code{firstActiveCol}}{firstActiveCol}

\item{\code{firstRow}}{firstRow}

\item{\code{firstCol}}{firstCol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_image"></a>}}
\if{latex}{\out{\hypertarget{method-add_image}{}}}
\subsection{Method \code{add_image()}}{
Insert an image into a sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$add_image(
  sheet,
  file,
  width = 6,
  height = 3,
  startRow = 1,
  startCol = 1,
  rowOffset = 0,
  colOffset = 0,
  units = "in",
  dpi = 300
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{file}}{file}

\item{\code{width}}{width}

\item{\code{height}}{height}

\item{\code{startRow}}{startRow}

\item{\code{startCol}}{startCol}

\item{\code{rowOffset}}{rowOffset}

\item{\code{colOffset}}{colOffset}

\item{\code{units}}{units}

\item{\code{dpi}}{dpi}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_plot"></a>}}
\if{latex}{\out{\hypertarget{method-add_plot}{}}}
\subsection{Method \code{add_plot()}}{
Add plot. A wrapper for add_image()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$add_plot(
  sheet,
  width = 6,
  height = 4,
  xy = NULL,
  startRow = 1,
  startCol = 1,
  rowOffset = 0,
  colOffset = 0,
  fileType = "png",
  units = "in",
  dpi = 300
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{width}}{width}

\item{\code{height}}{height}

\item{\code{xy}}{xy}

\item{\code{startRow}}{startRow}

\item{\code{startCol}}{startCol}

\item{\code{rowOffset}}{rowOffset}

\item{\code{colOffset}}{colOffset}

\item{\code{fileType}}{fileType}

\item{\code{units}}{units}

\item{\code{dpi}}{dpi}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints the \code{wbWorkbook} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{wbWorkbook} object, invisibly; called for its side-effects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-protect"></a>}}
\if{latex}{\out{\hypertarget{method-protect}{}}}
\subsection{Method \code{protect()}}{
Protect a workbook
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$protect(
  protect = TRUE,
  lockStructure = FALSE,
  lockWindows = FALSE,
  password = NULL,
  type = NULL,
  fileSharing = FALSE,
  username = unname(Sys.info()["user"]),
  readOnlyRecommended = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protect}}{protect}

\item{\code{lockStructure}}{lockStructure}

\item{\code{lockWindows}}{lockWindows}

\item{\code{password}}{password}

\item{\code{type}}{type}

\item{\code{fileSharing}}{fileSharing}

\item{\code{username}}{username}

\item{\code{readOnlyRecommended}}{readOnlyRecommended}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_creators"></a>}}
\if{latex}{\out{\hypertarget{method-set_creators}{}}}
\subsection{Method \code{set_creators()}}{
Set creator(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_creators(creators)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creators}}{A character vector of creators to set.  Duplicates are
ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_creators"></a>}}
\if{latex}{\out{\hypertarget{method-add_creators}{}}}
\subsection{Method \code{add_creators()}}{
Add creator(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$add_creators(creators)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creators}}{A character vector of creators to add.  Duplicates are
ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_creators"></a>}}
\if{latex}{\out{\hypertarget{method-remove_creators}{}}}
\subsection{Method \code{remove_creators()}}{
Remove creator(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_creators(creators)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creators}}{A character vector of creators to remove.  All duplicated
are removed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_last_modified_by"></a>}}
\if{latex}{\out{\hypertarget{method-set_last_modified_by}{}}}
\subsection{Method \code{set_last_modified_by()}}{
Change the last modified by
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_last_modified_by(LastModifiedBy = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LastModifiedBy}}{A new value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_setup"></a>}}
\if{latex}{\out{\hypertarget{method-page_setup}{}}}
\subsection{Method \code{page_setup()}}{
page_setup()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$page_setup(
  sheet,
  orientation = NULL,
  scale = 100,
  left = 0.7,
  right = 0.7,
  top = 0.75,
  bottom = 0.75,
  header = 0.3,
  footer = 0.3,
  fitToWidth = FALSE,
  fitToHeight = FALSE,
  paperSize = NULL,
  printTitleRows = NULL,
  printTitleCols = NULL,
  summaryRow = NULL,
  summaryCol = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{orientation}}{orientation}

\item{\code{scale}}{scale}

\item{\code{left}}{left}

\item{\code{right}}{right}

\item{\code{top}}{top}

\item{\code{bottom}}{bottom}

\item{\code{header}}{header}

\item{\code{footer}}{footer}

\item{\code{fitToWidth}}{fitToWidth}

\item{\code{fitToHeight}}{fitToHeight}

\item{\code{paperSize}}{paperSize}

\item{\code{printTitleRows}}{printTitleRows}

\item{\code{printTitleCols}}{printTitleCols}

\item{\code{summaryRow}}{summaryRow}

\item{\code{summaryCol}}{summaryCol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_header_footer"></a>}}
\if{latex}{\out{\hypertarget{method-set_header_footer}{}}}
\subsection{Method \code{set_header_footer()}}{
Sets headers and footers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_header_footer(
  sheet,
  header = NULL,
  footer = NULL,
  evenHeader = NULL,
  evenFooter = NULL,
  firstHeader = NULL,
  firstFooter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{header}}{header}

\item{\code{footer}}{footer}

\item{\code{evenHeader}}{evenHeader}

\item{\code{evenFooter}}{evenFooter}

\item{\code{firstHeader}}{firstHeader}

\item{\code{firstFooter}}{firstFooter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tables"></a>}}
\if{latex}{\out{\hypertarget{method-get_tables}{}}}
\subsection{Method \code{get_tables()}}{
get tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$get_tables(sheet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_tables"></a>}}
\if{latex}{\out{\hypertarget{method-remove_tables}{}}}
\subsection{Method \code{remove_tables()}}{
remove tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_tables(sheet, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{table}}{table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_filter"></a>}}
\if{latex}{\out{\hypertarget{method-add_filter}{}}}
\subsection{Method \code{add_filter()}}{
add filters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$add_filter(sheet, rows, cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{rows}}{rows}

\item{\code{cols}}{cols}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_filter"></a>}}
\if{latex}{\out{\hypertarget{method-remove_filter}{}}}
\subsection{Method \code{remove_filter()}}{
remove filters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_filter(sheet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grid_lines"></a>}}
\if{latex}{\out{\hypertarget{method-grid_lines}{}}}
\subsection{Method \code{grid_lines()}}{
grid lines
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$grid_lines(sheet, show = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{show}}{show}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_named_region"></a>}}
\if{latex}{\out{\hypertarget{method-add_named_region}{}}}
\subsection{Method \code{add_named_region()}}{
add a named region
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$add_named_region(
  sheet,
  cols,
  rows,
  name,
  localSheetId = NULL,
  overwrite = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{cols}}{cols}

\item{\code{rows}}{rows}

\item{\code{name}}{name}

\item{\code{localSheetId}}{localSheetId}

\item{\code{overwrite}}{overwrite}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_named_region"></a>}}
\if{latex}{\out{\hypertarget{method-remove_named_region}{}}}
\subsection{Method \code{remove_named_region()}}{
remove a named region
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$remove_named_region(sheet = NULL, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_order}{}}}
\subsection{Method \code{set_order()}}{
set worksheet order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_order(sheets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheets}}{sheets}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{wbWorkbook} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sheet_visibility"></a>}}
\if{latex}{\out{\hypertarget{method-get_sheet_visibility}{}}}
\subsection{Method \code{get_sheet_visibility()}}{
Get sheet visibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$get_sheet_visibility()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_sheet_visibility"></a>}}
\if{latex}{\out{\hypertarget{method-set_sheet_visibility}{}}}
\subsection{Method \code{set_sheet_visibility()}}{
Set sheet visibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$set_sheet_visibility(sheet, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{sheet}

\item{\code{value}}{value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wbWorkbook$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
