% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{writeData2}
\alias{writeData2}
\title{dummy function to write data}
\usage{
writeData2(
  wb,
  sheet,
  data,
  name = NULL,
  colNames = TRUE,
  rowNames = FALSE,
  startRow = 1,
  startCol = 1,
  removeCellStyle = FALSE
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet}

\item{data}{data to export}

\item{name}{If not NULL, a named region is defined.}

\item{colNames}{include colnames?}

\item{rowNames}{include rownames?}

\item{startRow}{row to place it}

\item{startCol}{col to place it}

\item{removeCellStyle}{keep the cell style?}
}
\description{
dummy function to write data
}
\details{
The string \code{"_openxlsx_NA"} is reserved for \code{openxlsx2}. If the data frame
contains this string, the output will be broken.
}
\examples{
# create a workbook and add some sheets
wb <- wb_workbook()

wb$add_worksheet("sheet1")
writeData2(wb, "sheet1", mtcars, colNames = TRUE, rowNames = TRUE)

wb$add_worksheet("sheet2")
writeData2(wb, "sheet2", cars, colNames = FALSE)

wb$add_worksheet("sheet3")
writeData2(wb, "sheet3", letters)

wb$add_worksheet("sheet4")
writeData2(wb, "sheet4", as.data.frame(Titanic), startRow = 2, startCol = 2)

\dontrun{
file <- tempfile(fileext = ".xlsx")
wb_save(wb, path = file, overwrite = TRUE)
file.remove(file)
}

}
