% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_node}
\alias{xml_node}
\title{xml_node}
\usage{
xml_node(
  xml,
  level1 = NULL,
  level2 = NULL,
  level3 = NULL,
  level4 = NULL,
  level5 = NULL,
  level6 = NULL
)
}
\arguments{
\item{xml}{something xml}

\item{level1}{to please check}

\item{level2}{to please check}

\item{level3}{to please check}

\item{level4}{to please check}

\item{level5}{to please check}

\item{level6}{to please check}
}
\description{
xml_node
}
\details{
This function returns XML nodes as used in openxlsx2. In theory they
could be returned as pointers as well, but this has not yet been implemented.
}
\examples{
  x <- read_xml("<a><b/></a>")
  # return a
  xml_node(x, "a")
  # return b. requires the path to the node
  xml_node(x, "a", "b")
}
