% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugixml.R
\name{xml_value}
\alias{xml_value}
\title{xml_value}
\usage{
xml_value(
  xml,
  level1 = NULL,
  level2 = NULL,
  level3 = NULL,
  level4 = NULL,
  level5 = NULL,
  level6 = NULL
)
}
\arguments{
\item{level1}{to please check}

\item{level2}{to please check}

\item{level3}{to please check}

\item{level4}{to please check}

\item{level5}{to please check}

\item{x}{something xml}
}
\description{
xml_value
}
